Variable selection with Permuted Inclusion Criterion

Permutation method to select variables for pair comparison data with Bradley-Terry model. This procedure consists of two steps. In the first step, btpermute adds to the set of original predictors an additional set of 'fake' variables generated by randomly permuting the original predictors. Permuted predictors are not expected to have any predictive power for pairwise rankings. In the second step, btpermute adds predictors to the Bradley-Terry model. Each predictors (real and permuted) is added to a null model separately and retained if the variable reduces the model deviance most strongly. This process is replicated n times with argument n.iterations, each time with a new random permutation that can be controlled for replicability using argument seed. Across the n n.iterations the function identifies the predictor that appeared most often as the most deviance-reducing one. When this has a real variable, it is added to the model. Variable selection stops when a permuted variable(s) is found to be the most frequent most deviance-reducing predictors.

btpermute(
  contests = NULL,
  predictors = NULL,
  n.iterations = 15,
  seed = NULL,
  ...
)

Arguments

contests
a data frame with pairwise binary contests with these variables 'id','player1','player2','win1','win2'; in that order. The id should match with those found in predictors

predictors
a data frame with player predictors with an id that should match with contests

n.iterations
integer, number of iterations to compute

seed
integer, the seed for random number generation. If NULL (the default), gosset will set the seed randomly

...
additional arguments passed to BradleyTerry2 methods

Value

an object of class gosset_btpermute with the final BTm() model, selected variables, seeds (random numbers) used for permutations and deviances

References

Lysen, S. (2009) Permuted inclusion criterion: A variable selection technique. University of Pennsylvania https://repository.upenn.edu/edissertations/28/

See also

Author

Jonathan Steinke and Kauê de Sousa

Examples

# \donttest{
require("BradleyTerry2")

data("kenyachoice", package = "gosset")

mod <- btpermute(contests = kenyachoice$contests,
                 predictors = kenyachoice$predictors,
                 n.iterations = 10,
                 seed = 1)

mod
# }