Gari-Eba Consumer Acceptability in Cameroon and Nigeria

Data from decentralized consumer trials of gari-eba, a sub-product of cassava (Manihot esculenta Crantz). Consumer testing was carried out in 2022 in Cameroon and Nigeria using the tricot approach. Diverse consumers in villages, towns and cities evaluated the overall acceptability of gari-eba made from 13 cassava genotypes. Apart from overall preference of the samples, the following traits were evaluated for eba based on the triangulated insights obtained by earlier survey and participatory work in the three areas. Nigeria (Osun and Benue States): colour, smoothness, mouldability, stretchability, and taste. Cameroon (Littoral zone): colour, odour, taste, firmness and stretchability. Traits in common are: colour, taste and stretchability.

cassava

Format

A data frame with 1000 records and 27 variables

id

The tricot package id.

option_a

The name of genotype A in the comparison.

option_b

The name of genotype B in the comparison.

option_c

The name of genotype C in the comparison.

country

The country where the experiment was conducted.

gender

The participant gender.

age

The participant age.

consumption

Indicates how often the participant consumes eba.

consumptionform

Indicates in which meal the participant consumes eba.

colour_pos

The sample ranked as best for colour ("A", "B" or "C").

colour_neg

The sample ranked as worst for colour ("A", "B" or "C").

smoothness_pos

The sample ranked as best for smoothness ("A", "B" or "C").

smoothness_neg

The sample ranked as worst for smoothness ("A", "B" or "C").

mouldability_pos

The sample ranked as best for mouldability ("A", "B" or "C").

mouldability_neg

The sample ranked as worst for mouldability ("A", "B" or "C").

stretchability_pos

The sample ranked as best for stretchability ("A", "B" or "C").

stretchability_neg

The sample ranked as worst for stretchability ("A", "B" or "C").

taste_pos

The sample ranked as best for taste ("A", "B" or "C").

taste_neg

The sample ranked as worst for taste ("A", "B" or "C").

odour_pos

The sample ranked as best for odour ("A", "B" or "C").

odour_neg

The sample ranked as worst for odour ("A", "B" or "C").

firmness_pos

The sample ranked as best for firmness ("A", "B" or "C").

firmness_neg

The sample ranked as worst for firmness ("A", "B" or "C").

overall_pos

The sample ranked as best overall ("A", "B" or "C").

overall_neg

The sample ranked as worst overall ("A", "B" or "C").

reason_like

Open question for the reason why the participant ranked the best sample as "best".

reason_dislike

Open question for the reason why the participant ranked the worst sample as "worst".

Source

Olaosebikan, et. al. (2023). Journal of the Science of Food and Agriculture. doi:10.1002/jsfa.12867