Cross-validation

Methods for measuring the performance of a predictive model on sets of test data.

crossvalidation(
  formula,
  data,
  k = 10,
  folds = NULL,
  mean.method = NULL,
  seed = NULL,
  ...
)

Arguments

formula
an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted, of the form \( y ~ x1 + ... + xn \)

data
a data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.

k
an integer for the number of splits in the cross-validation

folds
an optional vector specifying the folds in the cross-validation

mean.method
a character for the method to calculate the mean of cross-validation estimators. Options are: 'equal', an arithmetic mean; 'foldsize', a weighted mean by the size in each fold; 'stouffer' a weighted Z-test developed by Stouffer et al. (1949).

seed
The seed for random number generation. If NULL (the default), gosset will set the seed randomly

...
additional arguments passed to methods

Value

The cross-validation goodness-of-fit estimates, which are:

AIC

Akaike Information Criterion

deviance

Model deviance

logLik

Log-Likelihood

MaxLik

Maximum likelihood pseudo R-squared

CraggUhler

Cragg and Uhler's pseudo R-squared

Agresti

Agresti pseudo R-squared

Cross-validation estimates are computed using the trained models on the validation samples.
See also

Examples

# Generalized Linear Models

data("airquality")

crossvalidation(Temp ~ Wind + Solar.R,
                data = airquality,
                k = 3,
                family = poisson())

#> Model formula:
#> Temp ~ Wind + Solar.R 
#>  
#> Cross-validation estimates: 
#> # A tibble: 1 x 6
#>     AIC deviance logLik MaxLik CraggUhler Agresti
#>   <dbl>    <dbl>  <dbl>  <dbl>      <dbl>   <dbl>
#> 1  346.     340.  -170.  0.250      0.250  0.0420

                

########################################

# PlackettLuce Model
# beans data from PlackettLuce
library("PlackettLuce")

example("beans", package = "PlackettLuce")

#> 
#> beans> # Consider the best and worst rankings. These give the variety the
#> beans> # farmer thought was best or worst, coded as A, B or C for the
#> beans> # first, second or third variety assigned to the farmer
#> beans> # respectively.
#> beans> data(beans)
#> 
#> beans> head(beans[c("best", "worst")], 2)
#>   best worst
#> 1    C     A
#> 2    B     A
#> 
#> beans> # Fill in the missing item
#> beans> beans$middle <- complete(beans[c("best", "worst")],
#> beans+                          items = c("A", "B", "C"))
#> 
#> beans> head(beans[c("best", "middle", "worst")], 2)
#>   best middle worst
#> 1    C      B     A
#> 2    B      C     A
#> 
#> beans> # This gives an ordering of the three varieties the farmer was
#> beans> # given. The names of these varieties are stored in separate
#> beans> # columns
#> beans> varieties <- beans[c("variety_a", "variety_b", "variety_c")]
#> 
#> beans> head(varieties, 2)
#>     variety_a       variety_b   variety_c
#> 1 BRT 103-182      SJC 730-79 PM2 Don Rey
#> 2   INTA Rojo INTA Centro Sur INTA Sequia
#> 
#> beans> # Use these names to decode the orderings of order 3
#> beans> order3 <- decode(beans[c("best", "middle", "worst")],
#> beans+                  items = beans[c("variety_a", "variety_b", "variety_c")],
#> beans+                  code = c("A", "B", "C"))
#> 
#> beans> # Now consider the paired comparisons agains the local variety
#> beans> head(beans[c("var_a", "var_b", "var_c")], 2)
#>   var_a  var_b  var_c
#> 1 Worse  Worse Better
#> 2 Worse Better Better
#> 
#> beans> # Convert these results to a vector and get the corresponding trial variety
#> beans> outcome <- unlist(beans[c("var_a", "var_b", "var_c")])
#> 
#> beans> trial_variety <- unlist(beans[c("variety_a", "variety_b", "variety_c")])
#> 
#> beans> # Create a data frame of the implied orderings of order 2
#> beans> order2 <- data.frame(Winner = ifelse(outcome == "Worse",
#> beans+                                      "Local", trial_variety),
#> beans+                      Loser = ifelse(outcome == "Worse",
#> beans+                                     trial_variety, "Local"),
#> beans+                      stringsAsFactors = FALSE, row.names = NULL)
#> 
#> beans> head(order2, 2)
#>   Winner       Loser
#> 1  Local BRT 103-182
#> 2  Local   INTA Rojo
#> 
#> beans> # Finally combine the rankings of order 2 and order 3
#> beans> R <- rbind(as.rankings(order3, input = "ordering"),
#> beans+            as.rankings(order2, input = "ordering"))
#> 
#> beans> head(R)
#> [1] "PM2 Don Rey > SJC 730-79 > BRT 103-182"  
#> [2] "INTA Centro Sur > INTA Sequia > INTA ..."
#> [3] "INTA Ferroso > INTA Matagalpa > BRT  ..."
#> [4] "INTA Rojo > INTA Centro Sur > ALS 0532-6"
#> [5] "PM2 Don Rey > INTA Sequia > SJC 730-79"  
#> [6] "ALS 0532-6 > INTA Matagalpa > INTA Rojo" 
#> 
#> beans> tail(R)
#> [1] "INTA Sequia > Local"    "INTA Sequia > Local"    "BRT 103-182 > Local"   
#> [4] "Local > INTA Matagalpa" "Local > INTA Rojo"      "Local > SJC 730-79"    

G <- grouped_rankings(R, rep(seq_len(nrow(beans)), 4))

#> Warning: 'grouped_rankings' is deprecated.
#> Use 'group' instead.
#> See help("Deprecated") and help("PlackettLuce-deprecated").

beans <- cbind(G, beans)

# take seasons as bins
k <- length(unique(beans$season))
folds <- as.integer(as.factor(beans$season))

crossvalidation(G ~ maxTN,
                data = beans,
                k = k,
                folds = folds,
                minsize = 100)

#> Model formula:
#> G ~ maxTN 
#>  
#> Cross-validation estimates: 
#> # A tibble: 1 x 6
#>     AIC deviance logLik MaxLik CraggUhler Agresti
#>   <dbl>    <dbl>  <dbl>  <dbl>      <dbl>   <dbl>
#> 1 1940.    1898.  -949. 0.0228     0.0228 0.00560

########################################

# BradleyTerry Model
library("psychotree")

#> Loading required package: partykit

#> Loading required package: grid

#> Loading required package: libcoin

#> Loading required package: mvtnorm

#> Loading required package: psychotools

# Germany's Next Topmodel 2007 data from psychotree
data("Topmodel2007", package = "psychotree")

crossvalidation(preference ~ .,
                data = Topmodel2007,
                k = 5)

#> Model formula:
#> preference ~ . 
#>  
#> Cross-validation estimates: 
#> # A tibble: 1 x 6
#>     AIC deviance logLik MaxLik CraggUhler Agresti
#>   <dbl>    <dbl>  <dbl>  <dbl>      <dbl>   <dbl>
#> 1  759.     725.  -363.  0.118      0.158  0.0919