Favourability scores

Compute favourability scores

favourite(data = NULL, items = NULL, input = NULL, reorder = TRUE)

# S3 method for cmb_fvrt
plot(x, ...)

Arguments

data
a data frame with columns specified by items and input values

items
a data frame or index of data for the column(s) containing the item names

input
a data frame or index of data for the column(s) containing the values to be ranked

reorder
logical, if items should be reordered from higher favourability score to least favourability score

x
an object of class 'cmb_fvrt' for the plotting method. Generates a 'ggplot' object that can be passed to any ggplot2 method

...
additional arguments passed to methods. See details

Value

a data.frame with descriptive statistics

N

number of times the given item was evaluated

best

relative number of times (in percentage) the given item was ranked as first

worst

relative number of times (in percentage) the given item was ranked as worst

wins

relative number of times (in percentage) the given item wins against the others

fav_score

the favourability score, which is the difference between best and worst performance

Examples

data("breadwheat", package = "gosset")

favourite(data = breadwheat,
          items = c(1:3),
          input = c(18:19))

#> # A tibble: 16 x 6
#>    items         N  best worst  wins fav_score
#>    <chr>     <dbl> <dbl> <dbl> <dbl>     <dbl>
#>  1 K9107        93  87.1  3.23 0.919     83.9 
#>  2 CSW18        94  78.7  7.45 0.856     71.3 
#>  3 PBW343       96  72.9  8.33 0.823     64.6 
#>  4 DBW17        97  47.4 14.4  0.665     33.0 
#>  5 HD2985       91  44.0 22.0  0.610     22.0 
#>  6 DPW621-50    86  36.0 26.7  0.547      9.30
#>  7 HD2932       89  28.1 29.2  0.494     -1.12
#>  8 PBW502       95  21.1 44.2  0.384    -23.2 
#>  9 PBW550       96  14.6 41.7  0.365    -27.1 
#> 10 WR544        93  15.1 44.1  0.355    -29.0 
#> 11 RAJ4120      85  16.5 47.1  0.347    -30.6 
#> 12 HD2824       87  14.9 46.0  0.345    -31.0 
#> 13 HP1633       97  16.5 49.5  0.335    -33.0 
#> 14 HW2045       90  12.2 46.7  0.328    -34.4 
#> 15 HI1563       92  10.9 50    0.304    -39.1 
#> 16 K0307        98  14.3 54.1  0.301    -39.8