Forward stepwise regression for model selection

Forward selection is a type of stepwise regression which begins with an empty model and adds in variables one by one. In each step one variable is added when it makes a single best improvement to the model.

forward(formula, data, k = NULL, folds = NULL, select.by = NULL,
  akaike.weights = FALSE, ncores = NULL, packages = NULL, ...)

Arguments

formula
a symbolic description of the model, if set as \( y ~ . \) all variables in data are used

data
a data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.

k
an integer for the number of splits in the cross-validation

folds
an optional vector specifying the folds in the cross-validation

select.by
a character for the goodness-of-fit statistical parameter to select the models. Set as 'deviance' by default.

akaike.weights
an optional logical object for averaging the goodness of fit coefficients with Akaike weights

ncores
a integer for the number of cores to be used in the parallel computing. Set as 1 by default (no parallelisation)

packages
an optional character vector of packages that the parallel tasks depend on, ignore if all required packages are loaded to .GlobalEnv

...
additional arguments passed to methods

Value

The cross-validation goodness-of-fit estimates for the best model, which are:

AIC

Akaike Information Criterion

deviance

Model deviance

logLik

Log-Likelihood

MaxLik

Maximum likelihood pseudo R-squared

CraggUhler

Cragg and Uhler's pseudo R-squared

Agresti

Agresti pseudo R-squared

Cross-validation estimates are computed using the fitted models on the validation samples.
See also

Examples

# NOT RUN {
library("gnm")
library("doParallel")
library("abind")

data("airquality")

mod <- forward(Temp ~ .,
               data = airquality,
               k = 3,
               select.by = "AIC",
               family = poisson(link = "log"))

# }