Kendall rank correlation coefficient

Compute Kendall rank correlation coefficient between two objects. Kendall is a coefficient used in statistics to measure the ordinal association between two measured quantities. A tau test is a non-parametric hypothesis test for statistical dependence based on the tau coefficient. The 'kendallTau' function applies the "kendall" method from 'stats::cor' with some previous treatment in the data, such as converting floating numbers into ranks (from the higher being the first and negative being the last) and the possibility to remove zeros from incomplete ranks

kendallTau(x, y, ...)

# S3 method for default
kendallTau(x, y, ...)

# S3 method for matrix
kendallTau(x, y, ...)

# S3 method for data.frame
kendallTau(x, y, ...)

# S3 method for rankings
kendallTau(x, y, ...)

# S3 method for grouped_rankings
kendallTau(x, y, ...)

# S3 method for paircomp
kendallTau(x, y, ...)

Arguments

x
a numeric vector, matrix or data frame

y
a vector, matrix or data frame with compatible dimensions to x

...
further arguments afecting the Kendall tau produced. See details

Value

The Kendall correlation coefficient and the Effective N, which is the equivalent N needed if all items were compared to all items. Can be used for significance testing.

Details

null.rm logical, to remove zeros from x and y

References

Kendall M. G. (1938). Biometrika, 30(1–2), 81–93. https://doi.org/10.1093/biomet/30.1-2.81.

See also

Examples

# check the correlation between observed rankings 
# and the predicted rankings from PlackettLuce

library("PlackettLuce")

R <- matrix(c(1, 2, 4, 3,
              1, 4, 2, 3,
              1, 2, 4, 3,
              1, 2, 4, 3,
              1, 3, 4, 2,
              1, 4, 3, 2), nrow = 6, byrow = TRUE)
colnames(R) <- LETTERS[1:4]

G <- group(as.rankings(R), 1:6)

mod <- pltree(G ~ 1, data = G)

preds <- predict(mod)

k <- kendallTau(R, preds)

# also applies to a single observation in the matrix

k <- kendallTau(R[1,], preds[1,])