Pairwise preference

Compute pairwise preference matrix between comparisons

preference(data = NULL, items = NULL, input = NULL, ...)

# S3 method for cmb_pref
plot(x, ...)

Arguments

data
a data frame with columns specified by items and input values

items
a data frame or index of data for the column(s) containing the item names

input
a data frame or index of data for the column(s) containing the values to be ranked

...
additional arguments passed to methods. See details

x
an object of class 'cmb_pref' for the plotting method. Generates a 'ggplot' object that can be passed to any ggplot2 method

Value

a data.frame with pairwise comparisons

player1

the first player in the comparison

player2

the second player in the comparison

win

times player1 wins against player2

ncontest

number of contests between player1 and player2

preference

number of times player1 is preferred against player2

Examples

# breadwheat data

data("breadwheat", package = "gosset")

preference(data = breadwheat,
           items = c(1:3),
           input = c(18:19))

#> # A tibble: 240 x 5
#>    player1 player2     win ncontest preference
#>    <chr>   <chr>     <dbl>    <dbl>      <dbl>
#>  1 CSW18   DBW17        13       13      1    
#>  2 CSW18   DPW621-50     8       10      0.8  
#>  3 CSW18   HD2824        7        8      0.875
#>  4 CSW18   HD2932       13       13      1    
#>  5 CSW18   HD2985       10       12      0.833
#>  6 CSW18   HI1563       13       15      0.867
#>  7 CSW18   HP1633       15       15      1    
#>  8 CSW18   HW2045       13       15      0.867
#>  9 CSW18   K0307        14       14      1    
#> 10 CSW18   K9107         3       11      0.273
#> # ... with 230 more rows

# beans data where each observer compares 3 varieties
# randomly distributed

library("PlackettLuce")
data("beans", package = "PlackettLuce")

preference(data = beans,
           items = c(1:3),
           input = c(4:5))

#> # A tibble: 90 x 5
#>    player1     player2           win ncontest preference
#>    <chr>       <chr>           <dbl>    <dbl>      <dbl>
#>  1 ALS 0532-6  BRT 103-182        27       64      0.422
#>  2 ALS 0532-6  INTA Centro Sur    21       58      0.362
#>  3 ALS 0532-6  INTA Ferroso       29       53      0.547
#>  4 ALS 0532-6  INTA Matagalpa     29       54      0.537
#>  5 ALS 0532-6  INTA Precoz        27       55      0.491
#>  6 ALS 0532-6  INTA Rojo          34       63      0.540
#>  7 ALS 0532-6  INTA Sequia        17       53      0.321
#>  8 ALS 0532-6  PM2 Don Rey        27       62      0.435
#>  9 ALS 0532-6  SJC 730-79         27       60      0.45 
#> 10 BRT 103-182 ALS 0532-6         37       64      0.578
#> # ... with 80 more rows

# also include the local item
preference(data = beans,
           items = c(1:3),
           input = c(4:5),
           additional.rank = beans[c(6:8)])

#> # A tibble: 110 x 5
#>    player1    player2           win ncontest preference
#>    <chr>      <chr>           <dbl>    <dbl>      <dbl>
#>  1 ALS 0532-6 BRT 103-182        27       64      0.422
#>  2 ALS 0532-6 INTA Centro Sur    21       58      0.362
#>  3 ALS 0532-6 INTA Ferroso       29       53      0.547
#>  4 ALS 0532-6 INTA Matagalpa     29       54      0.537
#>  5 ALS 0532-6 INTA Precoz        27       55      0.491
#>  6 ALS 0532-6 INTA Rojo          34       63      0.540
#>  7 ALS 0532-6 INTA Sequia        17       53      0.321
#>  8 ALS 0532-6 Local             106      261      0.406
#>  9 ALS 0532-6 PM2 Don Rey        27       62      0.435
#> 10 ALS 0532-6 SJC 730-79         27       60      0.45 
#> # ... with 100 more rows