Rainfall indices

Compute rainfall indices over a timespan

rainfall(object, day.one = NULL, span = NULL, index = NULL, ...)

Arguments

object
a numeric vector of geographic coordinates (lonlat) or a matrix containing the precipitation data. When lonlat is used, the function makes a call to nasapower::get_power to fetch and concatenate environmental data from NASA POWER (https://power.larc.nasa.gov/)

day.one
a vector of class 'Date' for the starting date to capture the environmental data

span
an integer or a vector with integers for the duration of the timespan to be captured

index
optional, a character or a vector with characters for the indices to be computed. All indices are given by default

...
additional arguments passed to methods

Value

A dataframe with selected indices. Options are:

MLDS

maximum length of consecutive dry days (r < 1 mm)

MLWS

maximum length of consecutive wet days (r >= 1 mm)

R10mm

number of heavy precipitation days (10 >= r < 20 mm)

R20mm

number of very heavy precipitation days (r >= 20)

SDII

simple daily intensity index (mean of wet days / total rainfall)

Rx1day

maximum 1-day rainfall (mm)

Rx5day

maximum 5-day rainfall (mm)

Rtotal

total rainfall (mm) in wet days (R >= 1)

Examples

library("gosset")
library("nasapower")

# random geographic locations around bbox(11, 12, 55, 58)
lonlat <- data.frame(lon = runif(10, 11, 12),
                     lat = runif(10, 55, 58))

# random planting dates around 2018-05-15 and 2018-05-20
pdates <- as.integer(runif(10, 17666, 17670))
pdates <- as.Date(pdates, origin = "1970-01-01")

# calculate rainfall for the first 50 days after planting
rainfall(lonlat,
         day.one = pdates,
         span = 50)

#> # A tibble: 10 x 8
#>     MLDS  MLWS R10mm R20mm  SDII Rx1day Rx5day Rtotal
#>    <int> <int> <int> <int> <dbl>  <dbl>  <dbl>  <dbl>
#>  1    15     6     1     0  5.58  11.0   26.9    59.8
#>  2    14     3     0     0  2.89   6.83  11.6    29.8
#>  3    11     3     0     0  3.72   8.26  17.5    44.6
#>  4    12     3     0     0  3.47   9.69  11.8    29.0
#>  5    15     2     0     0  4.31   4.88   9.31   18.4
#>  6    13     3     1     0  3.65  11.6   14.9    38.6
#>  7    13     6     1     0  5.57  11.0   26.9    59.8
#>  8    13     2     0     0  4.19   4.88   9.31   18.4
#>  9    11     3     0     0  3.65   9.81  15.1    44.8
#> 10    14     3     0     0  4.88   7.77  20.2    50.0

# include the first 15 days before planting (residual precipitation)
rainfall(lonlat,
         day.one = pdates,
         span = 50,
         days.before = 15)

#> # A tibble: 10 x 8
#>     MLDS  MLWS R10mm R20mm  SDII Rx1day Rx5day Rtotal
#>    <int> <int> <int> <int> <dbl>  <dbl>  <dbl>  <dbl>
#>  1    24     6     1     0  5.40  11.0   26.9    61.7
#>  2    14     3     0     0  2.90   6.83  11.6    32.0
#>  3    11     3     0     0  3.94   9.36  17.5    58.2
#>  4    15     3     0     0  3.37   9.69  11.8    41.3
#>  5    15     2     0     0  3.78   4.88   9.31   22.7
#>  6    13     3     1     0  3.62  11.6   14.9    40.8
#>  7    24     6     1     0  5.24  11.0   26.9    65.1
#>  8    15     2     0     0  3.55   4.88   9.31   25.1
#>  9    11     3     1     0  3.92  10.4   15.1    58.8
#> 10    14     3     0     0  4.69   7.77  20.2    51.1