Binary rankings from pairwise contests

Binary comparisons from a ranking object. Ties are not taken into account, then they are added as NA's.

rank_binomial(object, drop.null = FALSE, disaggregate = FALSE)

Arguments

object

an object of class rankings, grouped_rankings or paircomp

drop.null

logical, an optional argument to remove null contests

disaggregate

logical, if TRUE binaries are disaggregated by individual contests

Value

A data.frame with binary rank of pairwise contests:

player1

a factor with n levels for the first player in the contests

player2

a factor with n levels (same as player1) for the second player in the contests

win1

number of times player1 wins against player2

win2

number of times player2 wins against player1

References

Turner H. & Firth D. (2012). Journal of Statistical Software, 48(9), 1–21. doi:10.18637/jss.v048.i09

See also

Author

Kauê de Sousa

Examples

library("PlackettLuce")

R <- matrix(c(1, 2, 0, 0,
              4, 1, 2, 3,
              2, 4, 3, 1,
              1, 2, 3, 0,
              2, 1, 1, 0,
              1, 0, 3, 2), nrow = 6, byrow = TRUE)
colnames(R) <- c("apple", "banana", "orange", "pear")

R <- as.rankings(R)

rank_binomial(R)

rank_binomial(R, disaggregate = TRUE)