Build Plackett-Luce rankings from tricot dataset

Create an object of class "rankings" from tricot data. Tricot stands for "triadic comparison of technologies". Is a methodology to carry out large agronomic field experiments allowing the comparison between many alternative technologies, in many different environments. Each participant evaluates a set of three randomised technologies from a larger set.

rank_tricot(
  data = NULL,
  items = NULL,
  input = NULL,
  additional.rank = NULL,
  group = FALSE,
  ...
)

Arguments

data
a data frame with columns specified by items and input values

items
a data frame or index of data for the column(s) containing the item names

input
a data frame or index of data for the column(s) containing the values to be ranked

additional.rank
optional, a data frame for the comparisons between tricot items and the local item

group
logical, if TRUE return an object of class "grouped_rankings"

...
additional arguments passed to methods. See details

Value

a PlackettLuce "rankings" object, which is a matrix of dense rankings

Details

full.output: logical, to return a list with a "rankings", a "grouped_rankings" and the ordered items

References

van Etten J., et al. (2019). Experimental Agriculture, 55(S1), 275–296. https://doi.org/10.1017/S0014479716000739.

See also

Examples

# using breadwheat data
data("breadwheat", package = "gosset")

# convert the tricot rankings from breadwheat data
# into a object of class 'rankings' from PlackettLuce
R <- rank_tricot(breadwheat,
                 items = c("variety_a","variety_b","variety_c"),
                 input = c("overall_best","overall_worst"))

############################################################

# beans data where each observer compares 3 varieties randomly distributed
# from a list of 11 and additionally compares these 3 varieties
# with their local variety
library("PlackettLuce")
data("beans", package = "PlackettLuce")

# first build rankings with only tricot items
# and return an object of class 'rankings'
R <- rank_tricot(data = beans,
                    items = c(1:3),
                    input = c(4:5))
head(R)

#> [1] "PM2 Don Rey > SJC 730-79 > BRT 103-182"  
#> [2] "INTA Centro Sur > INTA Sequia > INTA ..."
#> [3] "INTA Ferroso > INTA Matagalpa > BRT  ..."
#> [4] "INTA Rojo > INTA Centro Sur > ALS 0532-6"
#> [5] "PM2 Don Rey > INTA Sequia > SJC 730-79"  
#> [6] "ALS 0532-6 > INTA Matagalpa > INTA Rojo" 

############################################################

# pass the comparison with local item as an additional rankings, then
# each of the 3 varieties are compared separately with the local item
# and return an object of class grouped_rankings
G <- rank_tricot(data = beans,
                    items = c(1:3),
                    input = c(4:5),
                    additional.rank = beans[c(6:8)],
                    group = TRUE)

head(G)

#>                                                 1 
#>  "PM2 Don Rey > SJ ..., Local > BRT 103-182, ..." 
#>                                                 2 
#>    "INTA Centro Sur  ..., Local > INTA Rojo, ..." 
#>                                                 3 
#> "INTA Ferroso > I ..., INTA Ferroso > Local, ..." 
#>                                                 4 
#> "INTA Rojo > INTA ..., INTA Centro Sur  ..., ..." 
#>                                                 5 
#>  "PM2 Don Rey > IN ..., INTA Sequia > Local, ..." 
#>                                                 6 
#>   "ALS 0532-6 > INT ..., ALS 0532-6 > Local, ..."