Probability of outperforming a check

Measures the precision of estimated values, and the potential response to selection on those estimated values compared to a check

reliability(x, ...)

# Default S3 method
reliability(x, y = NULL, ...)

# S3 method for class 'PlackettLuce'
reliability(x, ref, ...)

# S3 method for class 'pltree'
reliability(x, ref, ...)

Arguments

x

a numeric vector, or an object of class PlackettLuce or pltree

...

additional arguments passed to methods

y

numeric, the reference value

ref

a character or integer for indexing the element containing reference values in x

Value

the reliability based on the worth parameters

References

Eskridge and Mumm (1992). Theoret. Appl. Genetics 84, 494–500 doi:10.1007/BF00229512 .

Author

Kauê de Sousa, David Brown, Jacob van Etten

Examples

# Case 1. vector example

x = c(9.5, 12, 12.3, 17)

y = 11.2

reliability(x, y)
#> [1] 0.4589372 0.5172414 0.5234043 0.6028369

# Case 2. PlackettLuce model

library("PlackettLuce") 

R = matrix(c(1, 2, 4, 3,
              4, 1, 2, 3,
              2, 3, 1, 4,
              4, 2, 3, 1,
              2, 1, 4, 3,
              1, 4, 3, 2), nrow = 6, byrow = TRUE)
colnames(R) = c("apple", "banana", "orange", "pear")

mod = PlackettLuce(R)

reliability(mod, ref = "orange")
#>      item reliability reliabilitySE  worth Zvalue Pr(>|z|)
#>     <chr>       <dbl>         <dbl>  <dbl>  <dbl>    <dbl>
#> 1:  apple      0.5760        0.1387 0.2516 0.4041   0.6861
#> 2: banana      0.6406        0.1266 0.3301 0.7888   0.4302
#> 3: orange      0.5000        0.1260 0.1852     NA       NA
#> 4:   pear      0.5572        0.1178 0.2331 0.3281   0.7428

# \donttest{
# Case 3. PlackettLuce tree

data("beans", package = "PlackettLuce")

G = rank_tricot(data = beans,
                 items = c(1:3),
                 input = c(4:5),
                 group = TRUE,
                 additional.rank = beans[c(6:8)])

pld = cbind(G, beans[,c("maxTN", "season", "lon")])

tree = pltree(G ~ maxTN + season + lon, data = pld)

reliability(tree, ref = "Local")
#>      node            item reliability reliabilitySE  worth  Zvalue Pr(>|z|)
#>     <int>           <chr>       <dbl>         <dbl>  <dbl>   <dbl>    <dbl>
#> 1:      3      ALS 0532-6      0.3071        0.0262 0.0726 -5.9109   0.0000
#> 2:      3     BRT 103-182      0.3339        0.0281 0.0822 -4.9215   0.0000
#> 3:      3 INTA Centro Sur      0.3512        0.0277 0.0887 -4.5363   0.0000
#> 4:      3    INTA Ferroso      0.3588        0.0276 0.0917 -4.3423   0.0000
#> 5:      3  INTA Matagalpa      0.3070        0.0259 0.0726 -5.9499   0.0000
#> ---                                                                        
#> 29:     5       INTA Rojo      0.4617        0.0315 0.0825 -1.1109   0.2666
#> 30:     5     INTA Sequia      0.5584        0.0311 0.1216  1.7235   0.0848
#> 31:     5           Local      0.5000        0.0132 0.0962      NA       NA
#> 32:     5     PM2 Don Rey      0.5009        0.0322 0.0965  0.0256   0.9796
#> 33:     5      SJC 730-79      0.4168        0.0290 0.0687 -2.5542   0.0106

# }