Combine R objects by rows

Combine R objects when number and names of columns do not match

rowbind(x, ...)

# Default S3 method
rowbind(x, y, ...)

# S3 method for class 'list'
rowbind(x, ...)

Arguments

x

a R object, typically a data.frame, matrix or list

...

additional arguments passed to methods

y

a matrix, a data.frame (or any other object that can be coerced to data.frame)

Value

a data.frame with the combined data

Examples

 
# two data frames
rowbind(airquality, mtcars)
#>     Ozone Solar.R Wind Temp Month Day  mpg cyl  disp  hp drat    wt  qsec vs am
#> 1      41     190  7.4   67     5   1   NA  NA    NA  NA   NA    NA    NA NA NA
#> 2      36     118  8.0   72     5   2   NA  NA    NA  NA   NA    NA    NA NA NA
#> 3      12     149 12.6   74     5   3   NA  NA    NA  NA   NA    NA    NA NA NA
#> 4      18     313 11.5   62     5   4   NA  NA    NA  NA   NA    NA    NA NA NA
#> 5      NA      NA 14.3   56     5   5   NA  NA    NA  NA   NA    NA    NA NA NA
#> 6      28      NA 14.9   66     5   6   NA  NA    NA  NA   NA    NA    NA NA NA
#> 7      23     299  8.6   65     5   7   NA  NA    NA  NA   NA    NA    NA NA NA
#> 8      19      99 13.8   59     5   8   NA  NA    NA  NA   NA    NA    NA NA NA
#> 9       8      19 20.1   61     5   9   NA  NA    NA  NA   NA    NA    NA NA NA
#> 10     NA     194  8.6   69     5  10   NA  NA    NA  NA   NA    NA    NA NA NA
#> 11      7      NA  6.9   74     5  11   NA  NA    NA  NA   NA    NA    NA NA NA
#> 12     16     256  9.7   69     5  12   NA  NA    NA  NA   NA    NA    NA NA NA
#> 13     11     290  9.2   66     5  13   NA  NA    NA  NA   NA    NA    NA NA NA
#> 14     14     274 10.9   68     5  14   NA  NA    NA  NA   NA    NA    NA NA NA
#> 15     18      65 13.2   58     5  15   NA  NA    NA  NA   NA    NA    NA NA NA
#> 16     14     334 11.5   64     5  16   NA  NA    NA  NA   NA    NA    NA NA NA
#> 17     34     307 12.0   66     5  17   NA  NA    NA  NA   NA    NA    NA NA NA
#> 18      6      78 18.4   57     5  18   NA  NA    NA  NA   NA    NA    NA NA NA
#> 19     30     322 11.5   68     5  19   NA  NA    NA  NA   NA    NA    NA NA NA
#> 20     11      44  9.7   62     5  20   NA  NA    NA  NA   NA    NA    NA NA NA
#> 21      1       8  9.7   59     5  21   NA  NA    NA  NA   NA    NA    NA NA NA
#> 22     11     320 16.6   73     5  22   NA  NA    NA  NA   NA    NA    NA NA NA
#> 23      4      25  9.7   61     5  23   NA  NA    NA  NA   NA    NA    NA NA NA
#> 24     32      92 12.0   61     5  24   NA  NA    NA  NA   NA    NA    NA NA NA
#> 25     NA      66 16.6   57     5  25   NA  NA    NA  NA   NA    NA    NA NA NA
#> 26     NA     266 14.9   58     5  26   NA  NA    NA  NA   NA    NA    NA NA NA
#> 27     NA      NA  8.0   57     5  27   NA  NA    NA  NA   NA    NA    NA NA NA
#> 28     23      13 12.0   67     5  28   NA  NA    NA  NA   NA    NA    NA NA NA
#> 29     45     252 14.9   81     5  29   NA  NA    NA  NA   NA    NA    NA NA NA
#> 30    115     223  5.7   79     5  30   NA  NA    NA  NA   NA    NA    NA NA NA
#> 31     37     279  7.4   76     5  31   NA  NA    NA  NA   NA    NA    NA NA NA
#> 32     NA     286  8.6   78     6   1   NA  NA    NA  NA   NA    NA    NA NA NA
#> 33     NA     287  9.7   74     6   2   NA  NA    NA  NA   NA    NA    NA NA NA
#> 34     NA     242 16.1   67     6   3   NA  NA    NA  NA   NA    NA    NA NA NA
#> 35     NA     186  9.2   84     6   4   NA  NA    NA  NA   NA    NA    NA NA NA
#> 36     NA     220  8.6   85     6   5   NA  NA    NA  NA   NA    NA    NA NA NA
#> 37     NA     264 14.3   79     6   6   NA  NA    NA  NA   NA    NA    NA NA NA
#> 38     29     127  9.7   82     6   7   NA  NA    NA  NA   NA    NA    NA NA NA
#> 39     NA     273  6.9   87     6   8   NA  NA    NA  NA   NA    NA    NA NA NA
#> 40     71     291 13.8   90     6   9   NA  NA    NA  NA   NA    NA    NA NA NA
#> 41     39     323 11.5   87     6  10   NA  NA    NA  NA   NA    NA    NA NA NA
#> 42     NA     259 10.9   93     6  11   NA  NA    NA  NA   NA    NA    NA NA NA
#> 43     NA     250  9.2   92     6  12   NA  NA    NA  NA   NA    NA    NA NA NA
#> 44     23     148  8.0   82     6  13   NA  NA    NA  NA   NA    NA    NA NA NA
#> 45     NA     332 13.8   80     6  14   NA  NA    NA  NA   NA    NA    NA NA NA
#> 46     NA     322 11.5   79     6  15   NA  NA    NA  NA   NA    NA    NA NA NA
#> 47     21     191 14.9   77     6  16   NA  NA    NA  NA   NA    NA    NA NA NA
#> 48     37     284 20.7   72     6  17   NA  NA    NA  NA   NA    NA    NA NA NA
#> 49     20      37  9.2   65     6  18   NA  NA    NA  NA   NA    NA    NA NA NA
#> 50     12     120 11.5   73     6  19   NA  NA    NA  NA   NA    NA    NA NA NA
#> 51     13     137 10.3   76     6  20   NA  NA    NA  NA   NA    NA    NA NA NA
#> 52     NA     150  6.3   77     6  21   NA  NA    NA  NA   NA    NA    NA NA NA
#> 53     NA      59  1.7   76     6  22   NA  NA    NA  NA   NA    NA    NA NA NA
#> 54     NA      91  4.6   76     6  23   NA  NA    NA  NA   NA    NA    NA NA NA
#> 55     NA     250  6.3   76     6  24   NA  NA    NA  NA   NA    NA    NA NA NA
#> 56     NA     135  8.0   75     6  25   NA  NA    NA  NA   NA    NA    NA NA NA
#> 57     NA     127  8.0   78     6  26   NA  NA    NA  NA   NA    NA    NA NA NA
#> 58     NA      47 10.3   73     6  27   NA  NA    NA  NA   NA    NA    NA NA NA
#> 59     NA      98 11.5   80     6  28   NA  NA    NA  NA   NA    NA    NA NA NA
#> 60     NA      31 14.9   77     6  29   NA  NA    NA  NA   NA    NA    NA NA NA
#> 61     NA     138  8.0   83     6  30   NA  NA    NA  NA   NA    NA    NA NA NA
#> 62    135     269  4.1   84     7   1   NA  NA    NA  NA   NA    NA    NA NA NA
#> 63     49     248  9.2   85     7   2   NA  NA    NA  NA   NA    NA    NA NA NA
#> 64     32     236  9.2   81     7   3   NA  NA    NA  NA   NA    NA    NA NA NA
#> 65     NA     101 10.9   84     7   4   NA  NA    NA  NA   NA    NA    NA NA NA
#> 66     64     175  4.6   83     7   5   NA  NA    NA  NA   NA    NA    NA NA NA
#> 67     40     314 10.9   83     7   6   NA  NA    NA  NA   NA    NA    NA NA NA
#> 68     77     276  5.1   88     7   7   NA  NA    NA  NA   NA    NA    NA NA NA
#> 69     97     267  6.3   92     7   8   NA  NA    NA  NA   NA    NA    NA NA NA
#> 70     97     272  5.7   92     7   9   NA  NA    NA  NA   NA    NA    NA NA NA
#> 71     85     175  7.4   89     7  10   NA  NA    NA  NA   NA    NA    NA NA NA
#> 72     NA     139  8.6   82     7  11   NA  NA    NA  NA   NA    NA    NA NA NA
#> 73     10     264 14.3   73     7  12   NA  NA    NA  NA   NA    NA    NA NA NA
#> 74     27     175 14.9   81     7  13   NA  NA    NA  NA   NA    NA    NA NA NA
#> 75     NA     291 14.9   91     7  14   NA  NA    NA  NA   NA    NA    NA NA NA
#> 76      7      48 14.3   80     7  15   NA  NA    NA  NA   NA    NA    NA NA NA
#> 77     48     260  6.9   81     7  16   NA  NA    NA  NA   NA    NA    NA NA NA
#> 78     35     274 10.3   82     7  17   NA  NA    NA  NA   NA    NA    NA NA NA
#> 79     61     285  6.3   84     7  18   NA  NA    NA  NA   NA    NA    NA NA NA
#> 80     79     187  5.1   87     7  19   NA  NA    NA  NA   NA    NA    NA NA NA
#> 81     63     220 11.5   85     7  20   NA  NA    NA  NA   NA    NA    NA NA NA
#> 82     16       7  6.9   74     7  21   NA  NA    NA  NA   NA    NA    NA NA NA
#> 83     NA     258  9.7   81     7  22   NA  NA    NA  NA   NA    NA    NA NA NA
#> 84     NA     295 11.5   82     7  23   NA  NA    NA  NA   NA    NA    NA NA NA
#> 85     80     294  8.6   86     7  24   NA  NA    NA  NA   NA    NA    NA NA NA
#> 86    108     223  8.0   85     7  25   NA  NA    NA  NA   NA    NA    NA NA NA
#> 87     20      81  8.6   82     7  26   NA  NA    NA  NA   NA    NA    NA NA NA
#> 88     52      82 12.0   86     7  27   NA  NA    NA  NA   NA    NA    NA NA NA
#> 89     82     213  7.4   88     7  28   NA  NA    NA  NA   NA    NA    NA NA NA
#> 90     50     275  7.4   86     7  29   NA  NA    NA  NA   NA    NA    NA NA NA
#> 91     64     253  7.4   83     7  30   NA  NA    NA  NA   NA    NA    NA NA NA
#> 92     59     254  9.2   81     7  31   NA  NA    NA  NA   NA    NA    NA NA NA
#> 93     39      83  6.9   81     8   1   NA  NA    NA  NA   NA    NA    NA NA NA
#> 94      9      24 13.8   81     8   2   NA  NA    NA  NA   NA    NA    NA NA NA
#> 95     16      77  7.4   82     8   3   NA  NA    NA  NA   NA    NA    NA NA NA
#> 96     78      NA  6.9   86     8   4   NA  NA    NA  NA   NA    NA    NA NA NA
#> 97     35      NA  7.4   85     8   5   NA  NA    NA  NA   NA    NA    NA NA NA
#> 98     66      NA  4.6   87     8   6   NA  NA    NA  NA   NA    NA    NA NA NA
#> 99    122     255  4.0   89     8   7   NA  NA    NA  NA   NA    NA    NA NA NA
#> 100    89     229 10.3   90     8   8   NA  NA    NA  NA   NA    NA    NA NA NA
#> 101   110     207  8.0   90     8   9   NA  NA    NA  NA   NA    NA    NA NA NA
#> 102    NA     222  8.6   92     8  10   NA  NA    NA  NA   NA    NA    NA NA NA
#> 103    NA     137 11.5   86     8  11   NA  NA    NA  NA   NA    NA    NA NA NA
#> 104    44     192 11.5   86     8  12   NA  NA    NA  NA   NA    NA    NA NA NA
#> 105    28     273 11.5   82     8  13   NA  NA    NA  NA   NA    NA    NA NA NA
#> 106    65     157  9.7   80     8  14   NA  NA    NA  NA   NA    NA    NA NA NA
#> 107    NA      64 11.5   79     8  15   NA  NA    NA  NA   NA    NA    NA NA NA
#> 108    22      71 10.3   77     8  16   NA  NA    NA  NA   NA    NA    NA NA NA
#> 109    59      51  6.3   79     8  17   NA  NA    NA  NA   NA    NA    NA NA NA
#> 110    23     115  7.4   76     8  18   NA  NA    NA  NA   NA    NA    NA NA NA
#> 111    31     244 10.9   78     8  19   NA  NA    NA  NA   NA    NA    NA NA NA
#> 112    44     190 10.3   78     8  20   NA  NA    NA  NA   NA    NA    NA NA NA
#> 113    21     259 15.5   77     8  21   NA  NA    NA  NA   NA    NA    NA NA NA
#> 114     9      36 14.3   72     8  22   NA  NA    NA  NA   NA    NA    NA NA NA
#> 115    NA     255 12.6   75     8  23   NA  NA    NA  NA   NA    NA    NA NA NA
#> 116    45     212  9.7   79     8  24   NA  NA    NA  NA   NA    NA    NA NA NA
#> 117   168     238  3.4   81     8  25   NA  NA    NA  NA   NA    NA    NA NA NA
#> 118    73     215  8.0   86     8  26   NA  NA    NA  NA   NA    NA    NA NA NA
#> 119    NA     153  5.7   88     8  27   NA  NA    NA  NA   NA    NA    NA NA NA
#> 120    76     203  9.7   97     8  28   NA  NA    NA  NA   NA    NA    NA NA NA
#> 121   118     225  2.3   94     8  29   NA  NA    NA  NA   NA    NA    NA NA NA
#> 122    84     237  6.3   96     8  30   NA  NA    NA  NA   NA    NA    NA NA NA
#> 123    85     188  6.3   94     8  31   NA  NA    NA  NA   NA    NA    NA NA NA
#> 124    96     167  6.9   91     9   1   NA  NA    NA  NA   NA    NA    NA NA NA
#> 125    78     197  5.1   92     9   2   NA  NA    NA  NA   NA    NA    NA NA NA
#> 126    73     183  2.8   93     9   3   NA  NA    NA  NA   NA    NA    NA NA NA
#> 127    91     189  4.6   93     9   4   NA  NA    NA  NA   NA    NA    NA NA NA
#> 128    47      95  7.4   87     9   5   NA  NA    NA  NA   NA    NA    NA NA NA
#> 129    32      92 15.5   84     9   6   NA  NA    NA  NA   NA    NA    NA NA NA
#> 130    20     252 10.9   80     9   7   NA  NA    NA  NA   NA    NA    NA NA NA
#> 131    23     220 10.3   78     9   8   NA  NA    NA  NA   NA    NA    NA NA NA
#> 132    21     230 10.9   75     9   9   NA  NA    NA  NA   NA    NA    NA NA NA
#> 133    24     259  9.7   73     9  10   NA  NA    NA  NA   NA    NA    NA NA NA
#> 134    44     236 14.9   81     9  11   NA  NA    NA  NA   NA    NA    NA NA NA
#> 135    21     259 15.5   76     9  12   NA  NA    NA  NA   NA    NA    NA NA NA
#> 136    28     238  6.3   77     9  13   NA  NA    NA  NA   NA    NA    NA NA NA
#> 137     9      24 10.9   71     9  14   NA  NA    NA  NA   NA    NA    NA NA NA
#> 138    13     112 11.5   71     9  15   NA  NA    NA  NA   NA    NA    NA NA NA
#> 139    46     237  6.9   78     9  16   NA  NA    NA  NA   NA    NA    NA NA NA
#> 140    18     224 13.8   67     9  17   NA  NA    NA  NA   NA    NA    NA NA NA
#> 141    13      27 10.3   76     9  18   NA  NA    NA  NA   NA    NA    NA NA NA
#> 142    24     238 10.3   68     9  19   NA  NA    NA  NA   NA    NA    NA NA NA
#> 143    16     201  8.0   82     9  20   NA  NA    NA  NA   NA    NA    NA NA NA
#> 144    13     238 12.6   64     9  21   NA  NA    NA  NA   NA    NA    NA NA NA
#> 145    23      14  9.2   71     9  22   NA  NA    NA  NA   NA    NA    NA NA NA
#> 146    36     139 10.3   81     9  23   NA  NA    NA  NA   NA    NA    NA NA NA
#> 147     7      49 10.3   69     9  24   NA  NA    NA  NA   NA    NA    NA NA NA
#> 148    14      20 16.6   63     9  25   NA  NA    NA  NA   NA    NA    NA NA NA
#> 149    30     193  6.9   70     9  26   NA  NA    NA  NA   NA    NA    NA NA NA
#> 150    NA     145 13.2   77     9  27   NA  NA    NA  NA   NA    NA    NA NA NA
#> 151    14     191 14.3   75     9  28   NA  NA    NA  NA   NA    NA    NA NA NA
#> 152    18     131  8.0   76     9  29   NA  NA    NA  NA   NA    NA    NA NA NA
#> 153    20     223 11.5   68     9  30   NA  NA    NA  NA   NA    NA    NA NA NA
#> 154    NA      NA   NA   NA    NA  NA 21.0   6 160.0 110 3.90 2.620 16.46  0  1
#> 155    NA      NA   NA   NA    NA  NA 21.0   6 160.0 110 3.90 2.875 17.02  0  1
#> 156    NA      NA   NA   NA    NA  NA 22.8   4 108.0  93 3.85 2.320 18.61  1  1
#> 157    NA      NA   NA   NA    NA  NA 21.4   6 258.0 110 3.08 3.215 19.44  1  0
#> 158    NA      NA   NA   NA    NA  NA 18.7   8 360.0 175 3.15 3.440 17.02  0  0
#> 159    NA      NA   NA   NA    NA  NA 18.1   6 225.0 105 2.76 3.460 20.22  1  0
#> 160    NA      NA   NA   NA    NA  NA 14.3   8 360.0 245 3.21 3.570 15.84  0  0
#> 161    NA      NA   NA   NA    NA  NA 24.4   4 146.7  62 3.69 3.190 20.00  1  0
#> 162    NA      NA   NA   NA    NA  NA 22.8   4 140.8  95 3.92 3.150 22.90  1  0
#> 163    NA      NA   NA   NA    NA  NA 19.2   6 167.6 123 3.92 3.440 18.30  1  0
#> 164    NA      NA   NA   NA    NA  NA 17.8   6 167.6 123 3.92 3.440 18.90  1  0
#> 165    NA      NA   NA   NA    NA  NA 16.4   8 275.8 180 3.07 4.070 17.40  0  0
#> 166    NA      NA   NA   NA    NA  NA 17.3   8 275.8 180 3.07 3.730 17.60  0  0
#> 167    NA      NA   NA   NA    NA  NA 15.2   8 275.8 180 3.07 3.780 18.00  0  0
#> 168    NA      NA   NA   NA    NA  NA 10.4   8 472.0 205 2.93 5.250 17.98  0  0
#> 169    NA      NA   NA   NA    NA  NA 10.4   8 460.0 215 3.00 5.424 17.82  0  0
#> 170    NA      NA   NA   NA    NA  NA 14.7   8 440.0 230 3.23 5.345 17.42  0  0
#> 171    NA      NA   NA   NA    NA  NA 32.4   4  78.7  66 4.08 2.200 19.47  1  1
#> 172    NA      NA   NA   NA    NA  NA 30.4   4  75.7  52 4.93 1.615 18.52  1  1
#> 173    NA      NA   NA   NA    NA  NA 33.9   4  71.1  65 4.22 1.835 19.90  1  1
#> 174    NA      NA   NA   NA    NA  NA 21.5   4 120.1  97 3.70 2.465 20.01  1  0
#> 175    NA      NA   NA   NA    NA  NA 15.5   8 318.0 150 2.76 3.520 16.87  0  0
#> 176    NA      NA   NA   NA    NA  NA 15.2   8 304.0 150 3.15 3.435 17.30  0  0
#> 177    NA      NA   NA   NA    NA  NA 13.3   8 350.0 245 3.73 3.840 15.41  0  0
#> 178    NA      NA   NA   NA    NA  NA 19.2   8 400.0 175 3.08 3.845 17.05  0  0
#> 179    NA      NA   NA   NA    NA  NA 27.3   4  79.0  66 4.08 1.935 18.90  1  1
#> 180    NA      NA   NA   NA    NA  NA 26.0   4 120.3  91 4.43 2.140 16.70  0  1
#> 181    NA      NA   NA   NA    NA  NA 30.4   4  95.1 113 3.77 1.513 16.90  1  1
#> 182    NA      NA   NA   NA    NA  NA 15.8   8 351.0 264 4.22 3.170 14.50  0  1
#> 183    NA      NA   NA   NA    NA  NA 19.7   6 145.0 175 3.62 2.770 15.50  0  1
#> 184    NA      NA   NA   NA    NA  NA 15.0   8 301.0 335 3.54 3.570 14.60  0  1
#> 185    NA      NA   NA   NA    NA  NA 21.4   4 121.0 109 4.11 2.780 18.60  1  1
#>     gear carb
#> 1     NA   NA
#> 2     NA   NA
#> 3     NA   NA
#> 4     NA   NA
#> 5     NA   NA
#> 6     NA   NA
#> 7     NA   NA
#> 8     NA   NA
#> 9     NA   NA
#> 10    NA   NA
#> 11    NA   NA
#> 12    NA   NA
#> 13    NA   NA
#> 14    NA   NA
#> 15    NA   NA
#> 16    NA   NA
#> 17    NA   NA
#> 18    NA   NA
#> 19    NA   NA
#> 20    NA   NA
#> 21    NA   NA
#> 22    NA   NA
#> 23    NA   NA
#> 24    NA   NA
#> 25    NA   NA
#> 26    NA   NA
#> 27    NA   NA
#> 28    NA   NA
#> 29    NA   NA
#> 30    NA   NA
#> 31    NA   NA
#> 32    NA   NA
#> 33    NA   NA
#> 34    NA   NA
#> 35    NA   NA
#> 36    NA   NA
#> 37    NA   NA
#> 38    NA   NA
#> 39    NA   NA
#> 40    NA   NA
#> 41    NA   NA
#> 42    NA   NA
#> 43    NA   NA
#> 44    NA   NA
#> 45    NA   NA
#> 46    NA   NA
#> 47    NA   NA
#> 48    NA   NA
#> 49    NA   NA
#> 50    NA   NA
#> 51    NA   NA
#> 52    NA   NA
#> 53    NA   NA
#> 54    NA   NA
#> 55    NA   NA
#> 56    NA   NA
#> 57    NA   NA
#> 58    NA   NA
#> 59    NA   NA
#> 60    NA   NA
#> 61    NA   NA
#> 62    NA   NA
#> 63    NA   NA
#> 64    NA   NA
#> 65    NA   NA
#> 66    NA   NA
#> 67    NA   NA
#> 68    NA   NA
#> 69    NA   NA
#> 70    NA   NA
#> 71    NA   NA
#> 72    NA   NA
#> 73    NA   NA
#> 74    NA   NA
#> 75    NA   NA
#> 76    NA   NA
#> 77    NA   NA
#> 78    NA   NA
#> 79    NA   NA
#> 80    NA   NA
#> 81    NA   NA
#> 82    NA   NA
#> 83    NA   NA
#> 84    NA   NA
#> 85    NA   NA
#> 86    NA   NA
#> 87    NA   NA
#> 88    NA   NA
#> 89    NA   NA
#> 90    NA   NA
#> 91    NA   NA
#> 92    NA   NA
#> 93    NA   NA
#> 94    NA   NA
#> 95    NA   NA
#> 96    NA   NA
#> 97    NA   NA
#> 98    NA   NA
#> 99    NA   NA
#> 100   NA   NA
#> 101   NA   NA
#> 102   NA   NA
#> 103   NA   NA
#> 104   NA   NA
#> 105   NA   NA
#> 106   NA   NA
#> 107   NA   NA
#> 108   NA   NA
#> 109   NA   NA
#> 110   NA   NA
#> 111   NA   NA
#> 112   NA   NA
#> 113   NA   NA
#> 114   NA   NA
#> 115   NA   NA
#> 116   NA   NA
#> 117   NA   NA
#> 118   NA   NA
#> 119   NA   NA
#> 120   NA   NA
#> 121   NA   NA
#> 122   NA   NA
#> 123   NA   NA
#> 124   NA   NA
#> 125   NA   NA
#> 126   NA   NA
#> 127   NA   NA
#> 128   NA   NA
#> 129   NA   NA
#> 130   NA   NA
#> 131   NA   NA
#> 132   NA   NA
#> 133   NA   NA
#> 134   NA   NA
#> 135   NA   NA
#> 136   NA   NA
#> 137   NA   NA
#> 138   NA   NA
#> 139   NA   NA
#> 140   NA   NA
#> 141   NA   NA
#> 142   NA   NA
#> 143   NA   NA
#> 144   NA   NA
#> 145   NA   NA
#> 146   NA   NA
#> 147   NA   NA
#> 148   NA   NA
#> 149   NA   NA
#> 150   NA   NA
#> 151   NA   NA
#> 152   NA   NA
#> 153   NA   NA
#> 154    4    4
#> 155    4    4
#> 156    4    1
#> 157    3    1
#> 158    3    2
#> 159    3    1
#> 160    3    4
#> 161    4    2
#> 162    4    2
#> 163    4    4
#> 164    4    4
#> 165    3    3
#> 166    3    3
#> 167    3    3
#> 168    3    4
#> 169    3    4
#> 170    3    4
#> 171    4    1
#> 172    4    2
#> 173    4    1
#> 174    3    1
#> 175    3    2
#> 176    3    2
#> 177    3    4
#> 178    3    2
#> 179    4    1
#> 180    5    2
#> 181    5    2
#> 182    5    4
#> 183    5    6
#> 184    5    8
#> 185    4    2

# a list of data frames
l = list(airquality, mtcars)
rowbind(l)
#>     Ozone Solar.R Wind Temp Month Day  mpg cyl  disp  hp drat    wt  qsec vs am
#> 1      41     190  7.4   67     5   1   NA  NA    NA  NA   NA    NA    NA NA NA
#> 2      36     118  8.0   72     5   2   NA  NA    NA  NA   NA    NA    NA NA NA
#> 3      12     149 12.6   74     5   3   NA  NA    NA  NA   NA    NA    NA NA NA
#> 4      18     313 11.5   62     5   4   NA  NA    NA  NA   NA    NA    NA NA NA
#> 5      NA      NA 14.3   56     5   5   NA  NA    NA  NA   NA    NA    NA NA NA
#> 6      28      NA 14.9   66     5   6   NA  NA    NA  NA   NA    NA    NA NA NA
#> 7      23     299  8.6   65     5   7   NA  NA    NA  NA   NA    NA    NA NA NA
#> 8      19      99 13.8   59     5   8   NA  NA    NA  NA   NA    NA    NA NA NA
#> 9       8      19 20.1   61     5   9   NA  NA    NA  NA   NA    NA    NA NA NA
#> 10     NA     194  8.6   69     5  10   NA  NA    NA  NA   NA    NA    NA NA NA
#> 11      7      NA  6.9   74     5  11   NA  NA    NA  NA   NA    NA    NA NA NA
#> 12     16     256  9.7   69     5  12   NA  NA    NA  NA   NA    NA    NA NA NA
#> 13     11     290  9.2   66     5  13   NA  NA    NA  NA   NA    NA    NA NA NA
#> 14     14     274 10.9   68     5  14   NA  NA    NA  NA   NA    NA    NA NA NA
#> 15     18      65 13.2   58     5  15   NA  NA    NA  NA   NA    NA    NA NA NA
#> 16     14     334 11.5   64     5  16   NA  NA    NA  NA   NA    NA    NA NA NA
#> 17     34     307 12.0   66     5  17   NA  NA    NA  NA   NA    NA    NA NA NA
#> 18      6      78 18.4   57     5  18   NA  NA    NA  NA   NA    NA    NA NA NA
#> 19     30     322 11.5   68     5  19   NA  NA    NA  NA   NA    NA    NA NA NA
#> 20     11      44  9.7   62     5  20   NA  NA    NA  NA   NA    NA    NA NA NA
#> 21      1       8  9.7   59     5  21   NA  NA    NA  NA   NA    NA    NA NA NA
#> 22     11     320 16.6   73     5  22   NA  NA    NA  NA   NA    NA    NA NA NA
#> 23      4      25  9.7   61     5  23   NA  NA    NA  NA   NA    NA    NA NA NA
#> 24     32      92 12.0   61     5  24   NA  NA    NA  NA   NA    NA    NA NA NA
#> 25     NA      66 16.6   57     5  25   NA  NA    NA  NA   NA    NA    NA NA NA
#> 26     NA     266 14.9   58     5  26   NA  NA    NA  NA   NA    NA    NA NA NA
#> 27     NA      NA  8.0   57     5  27   NA  NA    NA  NA   NA    NA    NA NA NA
#> 28     23      13 12.0   67     5  28   NA  NA    NA  NA   NA    NA    NA NA NA
#> 29     45     252 14.9   81     5  29   NA  NA    NA  NA   NA    NA    NA NA NA
#> 30    115     223  5.7   79     5  30   NA  NA    NA  NA   NA    NA    NA NA NA
#> 31     37     279  7.4   76     5  31   NA  NA    NA  NA   NA    NA    NA NA NA
#> 32     NA     286  8.6   78     6   1   NA  NA    NA  NA   NA    NA    NA NA NA
#> 33     NA     287  9.7   74     6   2   NA  NA    NA  NA   NA    NA    NA NA NA
#> 34     NA     242 16.1   67     6   3   NA  NA    NA  NA   NA    NA    NA NA NA
#> 35     NA     186  9.2   84     6   4   NA  NA    NA  NA   NA    NA    NA NA NA
#> 36     NA     220  8.6   85     6   5   NA  NA    NA  NA   NA    NA    NA NA NA
#> 37     NA     264 14.3   79     6   6   NA  NA    NA  NA   NA    NA    NA NA NA
#> 38     29     127  9.7   82     6   7   NA  NA    NA  NA   NA    NA    NA NA NA
#> 39     NA     273  6.9   87     6   8   NA  NA    NA  NA   NA    NA    NA NA NA
#> 40     71     291 13.8   90     6   9   NA  NA    NA  NA   NA    NA    NA NA NA
#> 41     39     323 11.5   87     6  10   NA  NA    NA  NA   NA    NA    NA NA NA
#> 42     NA     259 10.9   93     6  11   NA  NA    NA  NA   NA    NA    NA NA NA
#> 43     NA     250  9.2   92     6  12   NA  NA    NA  NA   NA    NA    NA NA NA
#> 44     23     148  8.0   82     6  13   NA  NA    NA  NA   NA    NA    NA NA NA
#> 45     NA     332 13.8   80     6  14   NA  NA    NA  NA   NA    NA    NA NA NA
#> 46     NA     322 11.5   79     6  15   NA  NA    NA  NA   NA    NA    NA NA NA
#> 47     21     191 14.9   77     6  16   NA  NA    NA  NA   NA    NA    NA NA NA
#> 48     37     284 20.7   72     6  17   NA  NA    NA  NA   NA    NA    NA NA NA
#> 49     20      37  9.2   65     6  18   NA  NA    NA  NA   NA    NA    NA NA NA
#> 50     12     120 11.5   73     6  19   NA  NA    NA  NA   NA    NA    NA NA NA
#> 51     13     137 10.3   76     6  20   NA  NA    NA  NA   NA    NA    NA NA NA
#> 52     NA     150  6.3   77     6  21   NA  NA    NA  NA   NA    NA    NA NA NA
#> 53     NA      59  1.7   76     6  22   NA  NA    NA  NA   NA    NA    NA NA NA
#> 54     NA      91  4.6   76     6  23   NA  NA    NA  NA   NA    NA    NA NA NA
#> 55     NA     250  6.3   76     6  24   NA  NA    NA  NA   NA    NA    NA NA NA
#> 56     NA     135  8.0   75     6  25   NA  NA    NA  NA   NA    NA    NA NA NA
#> 57     NA     127  8.0   78     6  26   NA  NA    NA  NA   NA    NA    NA NA NA
#> 58     NA      47 10.3   73     6  27   NA  NA    NA  NA   NA    NA    NA NA NA
#> 59     NA      98 11.5   80     6  28   NA  NA    NA  NA   NA    NA    NA NA NA
#> 60     NA      31 14.9   77     6  29   NA  NA    NA  NA   NA    NA    NA NA NA
#> 61     NA     138  8.0   83     6  30   NA  NA    NA  NA   NA    NA    NA NA NA
#> 62    135     269  4.1   84     7   1   NA  NA    NA  NA   NA    NA    NA NA NA
#> 63     49     248  9.2   85     7   2   NA  NA    NA  NA   NA    NA    NA NA NA
#> 64     32     236  9.2   81     7   3   NA  NA    NA  NA   NA    NA    NA NA NA
#> 65     NA     101 10.9   84     7   4   NA  NA    NA  NA   NA    NA    NA NA NA
#> 66     64     175  4.6   83     7   5   NA  NA    NA  NA   NA    NA    NA NA NA
#> 67     40     314 10.9   83     7   6   NA  NA    NA  NA   NA    NA    NA NA NA
#> 68     77     276  5.1   88     7   7   NA  NA    NA  NA   NA    NA    NA NA NA
#> 69     97     267  6.3   92     7   8   NA  NA    NA  NA   NA    NA    NA NA NA
#> 70     97     272  5.7   92     7   9   NA  NA    NA  NA   NA    NA    NA NA NA
#> 71     85     175  7.4   89     7  10   NA  NA    NA  NA   NA    NA    NA NA NA
#> 72     NA     139  8.6   82     7  11   NA  NA    NA  NA   NA    NA    NA NA NA
#> 73     10     264 14.3   73     7  12   NA  NA    NA  NA   NA    NA    NA NA NA
#> 74     27     175 14.9   81     7  13   NA  NA    NA  NA   NA    NA    NA NA NA
#> 75     NA     291 14.9   91     7  14   NA  NA    NA  NA   NA    NA    NA NA NA
#> 76      7      48 14.3   80     7  15   NA  NA    NA  NA   NA    NA    NA NA NA
#> 77     48     260  6.9   81     7  16   NA  NA    NA  NA   NA    NA    NA NA NA
#> 78     35     274 10.3   82     7  17   NA  NA    NA  NA   NA    NA    NA NA NA
#> 79     61     285  6.3   84     7  18   NA  NA    NA  NA   NA    NA    NA NA NA
#> 80     79     187  5.1   87     7  19   NA  NA    NA  NA   NA    NA    NA NA NA
#> 81     63     220 11.5   85     7  20   NA  NA    NA  NA   NA    NA    NA NA NA
#> 82     16       7  6.9   74     7  21   NA  NA    NA  NA   NA    NA    NA NA NA
#> 83     NA     258  9.7   81     7  22   NA  NA    NA  NA   NA    NA    NA NA NA
#> 84     NA     295 11.5   82     7  23   NA  NA    NA  NA   NA    NA    NA NA NA
#> 85     80     294  8.6   86     7  24   NA  NA    NA  NA   NA    NA    NA NA NA
#> 86    108     223  8.0   85     7  25   NA  NA    NA  NA   NA    NA    NA NA NA
#> 87     20      81  8.6   82     7  26   NA  NA    NA  NA   NA    NA    NA NA NA
#> 88     52      82 12.0   86     7  27   NA  NA    NA  NA   NA    NA    NA NA NA
#> 89     82     213  7.4   88     7  28   NA  NA    NA  NA   NA    NA    NA NA NA
#> 90     50     275  7.4   86     7  29   NA  NA    NA  NA   NA    NA    NA NA NA
#> 91     64     253  7.4   83     7  30   NA  NA    NA  NA   NA    NA    NA NA NA
#> 92     59     254  9.2   81     7  31   NA  NA    NA  NA   NA    NA    NA NA NA
#> 93     39      83  6.9   81     8   1   NA  NA    NA  NA   NA    NA    NA NA NA
#> 94      9      24 13.8   81     8   2   NA  NA    NA  NA   NA    NA    NA NA NA
#> 95     16      77  7.4   82     8   3   NA  NA    NA  NA   NA    NA    NA NA NA
#> 96     78      NA  6.9   86     8   4   NA  NA    NA  NA   NA    NA    NA NA NA
#> 97     35      NA  7.4   85     8   5   NA  NA    NA  NA   NA    NA    NA NA NA
#> 98     66      NA  4.6   87     8   6   NA  NA    NA  NA   NA    NA    NA NA NA
#> 99    122     255  4.0   89     8   7   NA  NA    NA  NA   NA    NA    NA NA NA
#> 100    89     229 10.3   90     8   8   NA  NA    NA  NA   NA    NA    NA NA NA
#> 101   110     207  8.0   90     8   9   NA  NA    NA  NA   NA    NA    NA NA NA
#> 102    NA     222  8.6   92     8  10   NA  NA    NA  NA   NA    NA    NA NA NA
#> 103    NA     137 11.5   86     8  11   NA  NA    NA  NA   NA    NA    NA NA NA
#> 104    44     192 11.5   86     8  12   NA  NA    NA  NA   NA    NA    NA NA NA
#> 105    28     273 11.5   82     8  13   NA  NA    NA  NA   NA    NA    NA NA NA
#> 106    65     157  9.7   80     8  14   NA  NA    NA  NA   NA    NA    NA NA NA
#> 107    NA      64 11.5   79     8  15   NA  NA    NA  NA   NA    NA    NA NA NA
#> 108    22      71 10.3   77     8  16   NA  NA    NA  NA   NA    NA    NA NA NA
#> 109    59      51  6.3   79     8  17   NA  NA    NA  NA   NA    NA    NA NA NA
#> 110    23     115  7.4   76     8  18   NA  NA    NA  NA   NA    NA    NA NA NA
#> 111    31     244 10.9   78     8  19   NA  NA    NA  NA   NA    NA    NA NA NA
#> 112    44     190 10.3   78     8  20   NA  NA    NA  NA   NA    NA    NA NA NA
#> 113    21     259 15.5   77     8  21   NA  NA    NA  NA   NA    NA    NA NA NA
#> 114     9      36 14.3   72     8  22   NA  NA    NA  NA   NA    NA    NA NA NA
#> 115    NA     255 12.6   75     8  23   NA  NA    NA  NA   NA    NA    NA NA NA
#> 116    45     212  9.7   79     8  24   NA  NA    NA  NA   NA    NA    NA NA NA
#> 117   168     238  3.4   81     8  25   NA  NA    NA  NA   NA    NA    NA NA NA
#> 118    73     215  8.0   86     8  26   NA  NA    NA  NA   NA    NA    NA NA NA
#> 119    NA     153  5.7   88     8  27   NA  NA    NA  NA   NA    NA    NA NA NA
#> 120    76     203  9.7   97     8  28   NA  NA    NA  NA   NA    NA    NA NA NA
#> 121   118     225  2.3   94     8  29   NA  NA    NA  NA   NA    NA    NA NA NA
#> 122    84     237  6.3   96     8  30   NA  NA    NA  NA   NA    NA    NA NA NA
#> 123    85     188  6.3   94     8  31   NA  NA    NA  NA   NA    NA    NA NA NA
#> 124    96     167  6.9   91     9   1   NA  NA    NA  NA   NA    NA    NA NA NA
#> 125    78     197  5.1   92     9   2   NA  NA    NA  NA   NA    NA    NA NA NA
#> 126    73     183  2.8   93     9   3   NA  NA    NA  NA   NA    NA    NA NA NA
#> 127    91     189  4.6   93     9   4   NA  NA    NA  NA   NA    NA    NA NA NA
#> 128    47      95  7.4   87     9   5   NA  NA    NA  NA   NA    NA    NA NA NA
#> 129    32      92 15.5   84     9   6   NA  NA    NA  NA   NA    NA    NA NA NA
#> 130    20     252 10.9   80     9   7   NA  NA    NA  NA   NA    NA    NA NA NA
#> 131    23     220 10.3   78     9   8   NA  NA    NA  NA   NA    NA    NA NA NA
#> 132    21     230 10.9   75     9   9   NA  NA    NA  NA   NA    NA    NA NA NA
#> 133    24     259  9.7   73     9  10   NA  NA    NA  NA   NA    NA    NA NA NA
#> 134    44     236 14.9   81     9  11   NA  NA    NA  NA   NA    NA    NA NA NA
#> 135    21     259 15.5   76     9  12   NA  NA    NA  NA   NA    NA    NA NA NA
#> 136    28     238  6.3   77     9  13   NA  NA    NA  NA   NA    NA    NA NA NA
#> 137     9      24 10.9   71     9  14   NA  NA    NA  NA   NA    NA    NA NA NA
#> 138    13     112 11.5   71     9  15   NA  NA    NA  NA   NA    NA    NA NA NA
#> 139    46     237  6.9   78     9  16   NA  NA    NA  NA   NA    NA    NA NA NA
#> 140    18     224 13.8   67     9  17   NA  NA    NA  NA   NA    NA    NA NA NA
#> 141    13      27 10.3   76     9  18   NA  NA    NA  NA   NA    NA    NA NA NA
#> 142    24     238 10.3   68     9  19   NA  NA    NA  NA   NA    NA    NA NA NA
#> 143    16     201  8.0   82     9  20   NA  NA    NA  NA   NA    NA    NA NA NA
#> 144    13     238 12.6   64     9  21   NA  NA    NA  NA   NA    NA    NA NA NA
#> 145    23      14  9.2   71     9  22   NA  NA    NA  NA   NA    NA    NA NA NA
#> 146    36     139 10.3   81     9  23   NA  NA    NA  NA   NA    NA    NA NA NA
#> 147     7      49 10.3   69     9  24   NA  NA    NA  NA   NA    NA    NA NA NA
#> 148    14      20 16.6   63     9  25   NA  NA    NA  NA   NA    NA    NA NA NA
#> 149    30     193  6.9   70     9  26   NA  NA    NA  NA   NA    NA    NA NA NA
#> 150    NA     145 13.2   77     9  27   NA  NA    NA  NA   NA    NA    NA NA NA
#> 151    14     191 14.3   75     9  28   NA  NA    NA  NA   NA    NA    NA NA NA
#> 152    18     131  8.0   76     9  29   NA  NA    NA  NA   NA    NA    NA NA NA
#> 153    20     223 11.5   68     9  30   NA  NA    NA  NA   NA    NA    NA NA NA
#> 154    NA      NA   NA   NA    NA  NA 21.0   6 160.0 110 3.90 2.620 16.46  0  1
#> 155    NA      NA   NA   NA    NA  NA 21.0   6 160.0 110 3.90 2.875 17.02  0  1
#> 156    NA      NA   NA   NA    NA  NA 22.8   4 108.0  93 3.85 2.320 18.61  1  1
#> 157    NA      NA   NA   NA    NA  NA 21.4   6 258.0 110 3.08 3.215 19.44  1  0
#> 158    NA      NA   NA   NA    NA  NA 18.7   8 360.0 175 3.15 3.440 17.02  0  0
#> 159    NA      NA   NA   NA    NA  NA 18.1   6 225.0 105 2.76 3.460 20.22  1  0
#> 160    NA      NA   NA   NA    NA  NA 14.3   8 360.0 245 3.21 3.570 15.84  0  0
#> 161    NA      NA   NA   NA    NA  NA 24.4   4 146.7  62 3.69 3.190 20.00  1  0
#> 162    NA      NA   NA   NA    NA  NA 22.8   4 140.8  95 3.92 3.150 22.90  1  0
#> 163    NA      NA   NA   NA    NA  NA 19.2   6 167.6 123 3.92 3.440 18.30  1  0
#> 164    NA      NA   NA   NA    NA  NA 17.8   6 167.6 123 3.92 3.440 18.90  1  0
#> 165    NA      NA   NA   NA    NA  NA 16.4   8 275.8 180 3.07 4.070 17.40  0  0
#> 166    NA      NA   NA   NA    NA  NA 17.3   8 275.8 180 3.07 3.730 17.60  0  0
#> 167    NA      NA   NA   NA    NA  NA 15.2   8 275.8 180 3.07 3.780 18.00  0  0
#> 168    NA      NA   NA   NA    NA  NA 10.4   8 472.0 205 2.93 5.250 17.98  0  0
#> 169    NA      NA   NA   NA    NA  NA 10.4   8 460.0 215 3.00 5.424 17.82  0  0
#> 170    NA      NA   NA   NA    NA  NA 14.7   8 440.0 230 3.23 5.345 17.42  0  0
#> 171    NA      NA   NA   NA    NA  NA 32.4   4  78.7  66 4.08 2.200 19.47  1  1
#> 172    NA      NA   NA   NA    NA  NA 30.4   4  75.7  52 4.93 1.615 18.52  1  1
#> 173    NA      NA   NA   NA    NA  NA 33.9   4  71.1  65 4.22 1.835 19.90  1  1
#> 174    NA      NA   NA   NA    NA  NA 21.5   4 120.1  97 3.70 2.465 20.01  1  0
#> 175    NA      NA   NA   NA    NA  NA 15.5   8 318.0 150 2.76 3.520 16.87  0  0
#> 176    NA      NA   NA   NA    NA  NA 15.2   8 304.0 150 3.15 3.435 17.30  0  0
#> 177    NA      NA   NA   NA    NA  NA 13.3   8 350.0 245 3.73 3.840 15.41  0  0
#> 178    NA      NA   NA   NA    NA  NA 19.2   8 400.0 175 3.08 3.845 17.05  0  0
#> 179    NA      NA   NA   NA    NA  NA 27.3   4  79.0  66 4.08 1.935 18.90  1  1
#> 180    NA      NA   NA   NA    NA  NA 26.0   4 120.3  91 4.43 2.140 16.70  0  1
#> 181    NA      NA   NA   NA    NA  NA 30.4   4  95.1 113 3.77 1.513 16.90  1  1
#> 182    NA      NA   NA   NA    NA  NA 15.8   8 351.0 264 4.22 3.170 14.50  0  1
#> 183    NA      NA   NA   NA    NA  NA 19.7   6 145.0 175 3.62 2.770 15.50  0  1
#> 184    NA      NA   NA   NA    NA  NA 15.0   8 301.0 335 3.54 3.570 14.60  0  1
#> 185    NA      NA   NA   NA    NA  NA 21.4   4 121.0 109 4.11 2.780 18.60  1  1
#>     gear carb
#> 1     NA   NA
#> 2     NA   NA
#> 3     NA   NA
#> 4     NA   NA
#> 5     NA   NA
#> 6     NA   NA
#> 7     NA   NA
#> 8     NA   NA
#> 9     NA   NA
#> 10    NA   NA
#> 11    NA   NA
#> 12    NA   NA
#> 13    NA   NA
#> 14    NA   NA
#> 15    NA   NA
#> 16    NA   NA
#> 17    NA   NA
#> 18    NA   NA
#> 19    NA   NA
#> 20    NA   NA
#> 21    NA   NA
#> 22    NA   NA
#> 23    NA   NA
#> 24    NA   NA
#> 25    NA   NA
#> 26    NA   NA
#> 27    NA   NA
#> 28    NA   NA
#> 29    NA   NA
#> 30    NA   NA
#> 31    NA   NA
#> 32    NA   NA
#> 33    NA   NA
#> 34    NA   NA
#> 35    NA   NA
#> 36    NA   NA
#> 37    NA   NA
#> 38    NA   NA
#> 39    NA   NA
#> 40    NA   NA
#> 41    NA   NA
#> 42    NA   NA
#> 43    NA   NA
#> 44    NA   NA
#> 45    NA   NA
#> 46    NA   NA
#> 47    NA   NA
#> 48    NA   NA
#> 49    NA   NA
#> 50    NA   NA
#> 51    NA   NA
#> 52    NA   NA
#> 53    NA   NA
#> 54    NA   NA
#> 55    NA   NA
#> 56    NA   NA
#> 57    NA   NA
#> 58    NA   NA
#> 59    NA   NA
#> 60    NA   NA
#> 61    NA   NA
#> 62    NA   NA
#> 63    NA   NA
#> 64    NA   NA
#> 65    NA   NA
#> 66    NA   NA
#> 67    NA   NA
#> 68    NA   NA
#> 69    NA   NA
#> 70    NA   NA
#> 71    NA   NA
#> 72    NA   NA
#> 73    NA   NA
#> 74    NA   NA
#> 75    NA   NA
#> 76    NA   NA
#> 77    NA   NA
#> 78    NA   NA
#> 79    NA   NA
#> 80    NA   NA
#> 81    NA   NA
#> 82    NA   NA
#> 83    NA   NA
#> 84    NA   NA
#> 85    NA   NA
#> 86    NA   NA
#> 87    NA   NA
#> 88    NA   NA
#> 89    NA   NA
#> 90    NA   NA
#> 91    NA   NA
#> 92    NA   NA
#> 93    NA   NA
#> 94    NA   NA
#> 95    NA   NA
#> 96    NA   NA
#> 97    NA   NA
#> 98    NA   NA
#> 99    NA   NA
#> 100   NA   NA
#> 101   NA   NA
#> 102   NA   NA
#> 103   NA   NA
#> 104   NA   NA
#> 105   NA   NA
#> 106   NA   NA
#> 107   NA   NA
#> 108   NA   NA
#> 109   NA   NA
#> 110   NA   NA
#> 111   NA   NA
#> 112   NA   NA
#> 113   NA   NA
#> 114   NA   NA
#> 115   NA   NA
#> 116   NA   NA
#> 117   NA   NA
#> 118   NA   NA
#> 119   NA   NA
#> 120   NA   NA
#> 121   NA   NA
#> 122   NA   NA
#> 123   NA   NA
#> 124   NA   NA
#> 125   NA   NA
#> 126   NA   NA
#> 127   NA   NA
#> 128   NA   NA
#> 129   NA   NA
#> 130   NA   NA
#> 131   NA   NA
#> 132   NA   NA
#> 133   NA   NA
#> 134   NA   NA
#> 135   NA   NA
#> 136   NA   NA
#> 137   NA   NA
#> 138   NA   NA
#> 139   NA   NA
#> 140   NA   NA
#> 141   NA   NA
#> 142   NA   NA
#> 143   NA   NA
#> 144   NA   NA
#> 145   NA   NA
#> 146   NA   NA
#> 147   NA   NA
#> 148   NA   NA
#> 149   NA   NA
#> 150   NA   NA
#> 151   NA   NA
#> 152   NA   NA
#> 153   NA   NA
#> 154    4    4
#> 155    4    4
#> 156    4    1
#> 157    3    1
#> 158    3    2
#> 159    3    1
#> 160    3    4
#> 161    4    2
#> 162    4    2
#> 163    4    4
#> 164    4    4
#> 165    3    3
#> 166    3    3
#> 167    3    3
#> 168    3    4
#> 169    3    4
#> 170    3    4
#> 171    4    1
#> 172    4    2
#> 173    4    1
#> 174    3    1
#> 175    3    2
#> 176    3    2
#> 177    3    4
#> 178    3    2
#> 179    4    1
#> 180    5    2
#> 181    5    2
#> 182    5    4
#> 183    5    6
#> 184    5    8
#> 185    4    2