Favourability scores

Summarise favourability scores from permutation contests

summarise_favourite(object, ...)

# S3 method for gosset_fvrt
plot(x, ...)

Arguments

object
an object of class rankings or grouped_rankings

...
additional arguments passed to methods. See details

x
an object of class gosset_fvrt for the plotting method. Generates a ggplot object that can be passed to any ggplot2 method

Value

A data.frame with the descriptive statistics:

N

number of times the given item was evaluated

best

relative number of times (in percentage) the given item was ranked as first

worst

relative number of times (in percentage) the given item was ranked as worst

wins

relative number of times (in percentage) the given item wins against the others

fav_score

the favourability score, which is the difference between best and worst performance

Details

minlength an integer, passed to abbreviate() to define the minimum length of the abbreviations

See also

Examples

data("breadwheat", package = "gosset")

R <- rank_tricot(data = breadwheat,
                 items = c("variety_a", "variety_b", "variety_c"),
                 input = c("overall_best", "overall_worst"))

fav <- summarise_favourite(R)

plot(fav)