Worst regret in regret-based models from recursive partitioning trees

Compute a worst regret index from recursive partitioning trees. Regret is an important heuristic in the behavioural sciences. Minimizing worst regret (the loss under the worst possible outcome) is a criterion that takes a conservative approach to risk analysis in diversification strategies.

worst_regret(object, ...)

Arguments

object
an object of class modelparty

...
further arguments passed to methods

Value

A data frame with probabilities of winning

items

the item names

win_probs

the probabilities of winning

worst_regret

the worst regret index

References

Loomes G. & Sugden R. (1982). The Economic Journal, 92(368), 805. https://doi.org/10.2307/2232669.

Bleichrodt H. & Wakker P. P. (2015). The Economic Journal, 125(583), 493–532. https://doi.org/10.1111/ecoj.12200.

Examples

library("PlackettLuce")
data("breadwheat", package = "gosset")

# convert the tricot rankings from breadwheat data
# into a object of class 'grouped_rankings'

G <- rank_tricot(breadwheat,
                 items = c("variety_a","variety_b","variety_c"),
                 input = c("overall_best","overall_worst"),
                 group = TRUE)

# combine grouped rankings with temperature indices
mydata <- cbind(G, breadwheat[c("lon","lat")])

# fit a pltree model using geographic data
mod <- pltree(G ~ ., data = mydata)

worst_regret(mod)

#> # A tibble: 16 x 3
#>    items     win_probs worst_regret
#>    <chr>         <dbl>        <dbl>
#>  1 K9107        0.363         0    
#>  2 CSW18        0.198         0.196
#>  3 PBW343       0.176         0.273
#>  4 DBW17        0.0606        0.303
#>  5 HD2985       0.0369        0.327
#>  6 DPW621-50    0.0341        0.338
#>  7 HD2932       0.0214        0.342
#>  8 RAJ4120      0.0128        0.350
#>  9 PBW550       0.0140        0.351
#> 10 HP1633       0.0122        0.352
#> 11 HW2045       0.0117        0.353
#> 12 HD2824       0.0112        0.353
#> 13 HI1563       0.0111        0.354
#> 14 PBW502       0.0142        0.355
#> 15 WR544        0.0122        0.361
#> 16 K0307        0.0111        0.363