Evapotranspiration

Compute evapotranspiration using the Blaney-Criddle method

ETo(object, day.one = NULL, span = NULL, lat = NULL, Kc = 1, p = NULL)

Arguments

object
a numeric vector of geographic coordinates (lonlat) or an array with two dimensions containing the temperature data; 1st dimension contains the day temperature and 2nd dimension the night temperature. When lonlat is used, the function makes a call to nasapower::get_power to fetch and concatenate environmental data from NASA POWER (https://power.larc.nasa.gov/)

day.one
a vector of class 'Date' for the starting date to capture the environmental data

span
an integer or a vector with integers for the duration of the timespan to be captured

lat
a vector for the latitude (in Decimal degrees)

Kc
a numeric value for the crop factor for water requirement

p
optional, a numeric value (from 0 to 1) used if lat is not given, representing the mean daily percentage of annual daytime hours for different latitudes

Value

The evapotranspiration in mm/day

Examples

library("gosset")
library("nasapower")

# random geographic locations around bbox(11, 12, 55, 58)
lonlat <- data.frame(lon = runif(10, 11, 12),
                     lat = runif(10, 55, 58))

# random planting dates around 2018-05-15 and 2018-05-20
pdates <- as.integer(runif(10, 17666, 17670))
pdates <- as.Date(pdates, origin = "1970-01-01")

# the evapotranspiration in the first 50 days after planting
ETo(lonlat,
    day.one = pdates,
    span = 50,
    lat = lonlat[["lat"]])

#> # A tibble: 10 x 1
#>      ETo
#>    <dbl>
#>  1  5.74
#>  2  5.68
#>  3  5.74
#>  4  5.78
#>  5  5.76
#>  6  5.69
#>  7  5.68
#>  8  5.73
#>  9  5.70
#> 10  6.04