Preferred bread wheat varieties

Data from decentralized on-farm trials of bread wheat (Triticum aestivum L.) varieties in Vaishali, India over the 2014's Rabi season. Farmers were asked to test three varieties of bread wheat randomly assigned as incomplete blocks of three varieties (out of 16 varieties) and assess which one had the best and worst performance considering four traits, germination, grain quality, yield and overall performance.

breadwheat

Format

A data frame with 493 records and 19 variables:

variety_a

The name of variety A in the comparison.

variety_b

The name of variety B in the comparison.

variety_c

The name of variety C in the comparison.

district

The administrative region where the experiment was stablished.

village

The village within the administrative region where the experiment was stablished.

participant_name

The participant name (ommited for protection and privacy).

age

The participant age.

gender

The participant gender M = Male; F = Female.

planting_date

The date which the experiment started.

lon

The longitude in which the experiment was stablished.

lat

The latitude in which the experiment was stablished.

germination_best

The variety ranked as best for germination ("A", "B" or "C").

germination_worst

The variety ranked as worst for germination ("A", "B" or "C").

grainquality_best

The variety ranked as best for grain quality ("A", "B" or "C").

grainquality_worst

The variety ranked as worst for grain quality ("A", "B" or "C").

yield_best

The variety ranked as best for yield ("A", "B" or "C").

yield_worst

The variety ranked as worst for yield ("A", "B" or "C").

overall_best

The variety ranked as best for overall perfomance ("A", "B" or "C").

overall_worst

The variety ranked as worst for overall perfomance ("A", "B" or "C").

Source

The data were provided by Bioversity International https://www.bioversityinternational.org/, a CGIAR Research Centre, through the ClimMob platform https://climmob.net.