Cross-validation

Methods for measuring the performance of a predictive model on sets of test data in Bradley-Terry, Generalized Linear, Generalized Nonlinear or Plackett-Luce models.

crossvalidation(
  formula,
  data,
  k = 10,
  folds = NULL,
  mean.method = NULL,
  seed = NULL,
  ...
)

Arguments

formula
an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted, of the form \(y ~ x1 + ... + xn\)

data
a data frame (or object coercible by as.data.frame to a data frame) containing the variables in the model

k
an integer for the number of bins in the cross-validation

folds
an optional vector specifying the folds in the cross-validation

mean.method
a character for the method to calculate the mean of cross-validation estimators. Options are: 'equal', arithmetic mean; 'foldsize', weighted mean by the size in each fold; 'stouffer' weighted through Z-test. See references

seed
The seed for random number generation. If NULL (the default), gosset will set the seed randomly

...
additional arguments passed the methods of the chosen model

Value

The cross-validation goodness-of-fit estimates, which are:

AIC

Akaike Information Criterion

deviance

Model deviance

logLik

Log-Likelihood

MaxLik

Maximum likelihood pseudo R-squared

CraggUhler

Cragg and Uhler's pseudo R-squared

Agresti

Agresti pseudo R-squared

kendallTau

the Kendall correlation coefficient, only for Plackett-Luce models

References

Agresti A. (2002). Categorical Data Analysis. http://doi.wiley.com/10.1002/0471249688

Elder J. F. (2003). Journal of Computational and Graphical Statistics, 12(4), 853–864. https://doi.org/10.1198/1061860032733.

James G., et al. (2013). An Introduction to Statistical Learning: with Applications in R. https://doi.org/10.1007/978-1-4614-7138-7

Whitlock M. C. (2005). Journal of Evolutionary Biology, 18(5), 1368–1373. https://doi.org/10.1111/j.1420-9101.2005.00917.x.

See also

Examples

# Generalized Linear Models

data("airquality")

cv <- crossvalidation(Temp ~ Wind + Solar.R,
                      data = airquality,
                      k = 3,
                      seed = 999,
                      family = poisson())

# \donttest{
# Plackett-Luce Model
# beans data from PlackettLuce
library("PlackettLuce")
data("beans", package = "PlackettLuce")

G <- rank_tricot(data = beans,
                 items = c(1:3),
                 input = c(4:5),
                 additional.rank = beans[c(6:8)],
                 group = TRUE)

beans <- cbind(G, beans)

# take seasons as bins
k <- length(unique(beans$season))
folds <- as.integer(as.factor(beans$season))

cv <- crossvalidation(G ~ maxTN,
                      data = beans,
                      k = k,
                      folds = folds,
                      minsize = 100)

########################################

# Bradley-Terry Model
library("psychotree")
# Germany's Next Topmodel 2007 data from psychotree
data("Topmodel2007", package = "psychotree")

cv <- crossvalidation(preference ~ .,
                      data = Topmodel2007,
                      k = 5)

# }