Forward stepwise regression for model selection

Forward selection is a type of stepwise regression which begins with an empty model and adds in variables one by one. In each step one variable is added when it makes a single best improvement to the model.

forward(
  formula,
  data,
  k = NULL,
  folds = NULL,
  select.by = NULL,
  akaike.weights = FALSE,
  ncores = NULL,
  packages = NULL,
  seed = NULL,
  ...
)

Arguments

formula
a symbolic description of the model, if set as \( y ~ . \) all variables in data are used

data
a data frame (or object coercible by as.data.frame to a data frame) containing the variables in the model

k
an integer for the number of bins in the cross-validation

folds
an optional vector specifying the folds in the cross-validation

select.by
a character for the goodness-of-fit statistical parameter to select the models. Set as 'deviance' by default. Check 'Value' for details in outputs and goodness-of-fit options.

akaike.weights
optional, logical object for averaging the goodness-of-fit coefficients with Akaike weights

ncores
an integer for the number of cores to be used in the parallel computing. Set as 1 by default (no parallelisation)

packages
character vector of packages that the parallel tasks depend on. If you use GLM or GNM you should call for gnm, for Plackett-Luce, use PlackettLuce, for Bradley-Terry, use psychotree

seed
integer, the seed for random number generation. If NULL (the default), gosset will set the seed randomly

...
additional arguments passed the methods of the chosen model

Value

an object of class gosset_cv with the cross-validation goodness-of-fit estimates, which are:

AIC

Akaike Information Criterion

deviance

Model deviance

logLik

Log-Likelihood

MaxLik

Maximum likelihood pseudo R-squared

CraggUhler

Cragg and Uhler's pseudo R-squared

Agresti

Agresti pseudo R-squared

Cross-validation estimates are computed using the fitted models on the validation samples.
See also

Examples

require("gnm")
require("foreach")
require("abind")
require("doParallel")

data("airquality")

mod <- forward(Temp ~ .,
               data = airquality,
               k = 3,
               select.by = "AIC",
               packages = "gnm",
               family = poisson(link = "log"))

# \donttest{
##########################################
# Bradley-Terry Model
require("psychotree")

# Germany's Next Topmodel 2007 data from psychotree
data("Topmodel2007", package = "psychotree")

forward(preference ~ .,
        data = Topmodel2007,
        k = 5,
        packages = "psychotree",
        alpha = 0.01,
        minsize = 30)

# }