Kendall rank correlation coefficient

Compute Kendall rank correlation coefficient between two objects. Kendall is a coefficient used in statistics to measure the ordinal association between two measured quantities. A tau test is a non-parametric hypothesis test for statistical dependence based on the tau coefficient. The 'kendallTau' function applies the "kendall" method from 'stats::cor' with some previous treatment in the data, such as converting floating numbers into ranks (from the higher being the first and negative being the last) and the possibility to remove zeros from incomplete ranks

kendallTau(x, y, null.rm = TRUE, average = TRUE, na.omit = FALSE, ...)

# Default S3 method
kendallTau(x, y, null.rm = TRUE, ...)

# S3 method for class 'matrix'
kendallTau(x, y, null.rm = TRUE, average = TRUE, na.omit = FALSE, ...)

# S3 method for class 'rankings'
kendallTau(x, y, ...)

# S3 method for class 'grouped_rankings'
kendallTau(x, y, ...)

# S3 method for class 'paircomp'
kendallTau(x, y, ...)

Arguments

x

a numeric vector, matrix or data frame

y

a vector, matrix or data frame with compatible dimensions to x

null.rm

logical, to remove zeros from x and y

average

logical, if FALSE returns the kendall and N-effective for each entry

na.omit

logical, if TRUE ignores entries with kendall = NA when computing the average

...

further arguments affecting the Kendall tau produced. See details

Value

The Kendall correlation coefficient and the Effective N, which is the equivalent N needed if all items were compared to all items. Can be used for significance testing.

References

Kendall M. G. (1938). Biometrika, 30(1–2), 81–93. doi:10.1093/biomet/30.1-2.81

See also

Author

Kauê de Sousa and Jacob van Etten

Examples

# Vector based example same as stats::cor(x, y, method = "kendall")
# but showing N-effective
x = c(1, 2, 3, 4, 5)

y = c(1, 1, 3, 2, NA)

w = c(1, 1, 3, 2, 5)

kendallTau(x, y)
#>    kendallTau N_effective Zvalue Pr(>|z|)
#>         <dbl>       <dbl>  <dbl>    <dbl>
#> 1:     0.5477      4.0000 1.1163   0.1321

kendallTau(x, w)
#>    kendallTau N_effective Zvalue Pr(>|z|)
#>         <dbl>       <dbl>  <dbl>    <dbl>
#> 1:     0.7379      5.0000 1.8074   0.0354

# Matrix and PlacketLuce ranking example 

library("PlackettLuce")
 
R = matrix(c(1, 2, 4, 3,
             1, 4, 2, 3,
             1, 2, NA, 3,
             1, 2, 4, 3,
             1, 3, 4, 2,
             1, 4, 3, 2), nrow = 6, byrow = TRUE)
colnames(R) = LETTERS[1:4]

G = group(as.rankings(R), 1:6)

mod = pltree(G ~ 1, data = G)

preds = predict(mod)

kendallTau(R, preds)
#>    kendallTau N_effective Zvalue Pr(>|z|)
#>         <dbl>       <dbl>  <dbl>    <dbl>
#> 1:     0.6232      7.3007 2.0252   0.0214

# Also returns raw values (no average) 

kendallTau(R, preds, average = FALSE)
#>    kendallTau N_effective Zvalue Pr(>|z|)
#>         <dbl>       <dbl>  <dbl>    <dbl>
#> 1:     0.6667      4.0000 1.3587   0.0871
#> 2:     0.3333      4.0000 0.6794   0.2485
#> 3:     0.3333      3.0000 0.5222   0.3008
#> 4:     0.6667      4.0000 1.3587   0.0871
#> 5:     1.0000      4.0000 2.0381   0.0208
#> 6:     0.6667      4.0000 1.3587   0.0871

# Choose to ignore entries with NA
R2 = matrix(c(1, 2, 4, 3,
              1, 4, 2, 3,
              NA, NA, NA, NA,
              1, 2, 4, 3,
              1, 3, 4, 2,
              1, 4, 3, 2), nrow = 6, byrow = TRUE)

kendallTau(R, R2, average = FALSE)
#>    kendallTau N_effective Zvalue Pr(>|z|)
#>         <dbl>       <dbl>  <dbl>    <dbl>
#> 1:     1.0000      4.0000 2.0381   0.0208
#> 2:     1.0000      4.0000 2.0381   0.0208
#> 3:         NA      1.0000     NA       NA
#> 4:     1.0000      4.0000 2.0381   0.0208
#> 5:     1.0000      4.0000 2.0381   0.0208
#> 6:     1.0000      4.0000 2.0381   0.0208

kendallTau(R, R2, average = TRUE)
#>    kendallTau N_effective Zvalue Pr(>|z|)
#>         <dbl>       <dbl>  <dbl>    <dbl>
#> 1:     0.9524      7.0000 3.0038   0.0013

kendallTau(R, R2, average = TRUE, na.omit = TRUE)
#>    kendallTau N_effective Zvalue Pr(>|z|)
#>         <dbl>       <dbl>  <dbl>    <dbl>
#> 1:     1.0000      6.8443 3.1035   0.0010