Regret-based values for risk assessment

Regret is an important heuristic in the behavioural sciences. Minimizing worst regret (the loss under the worst possible outcome) is a criterion that takes a conservative approach to risk analysis in diversification strategies.

regret(object, ..., bootstrap = TRUE, normalize = TRUE)

# S3 method for default
regret(object, ..., values, items, group, bootstrap = TRUE, normalize = TRUE)

# S3 method for pltree
regret(object, bootstrap = TRUE, normalize = TRUE, ...)

# S3 method for list
regret(object, bootstrap = TRUE, normalize = TRUE, ...)

Arguments

object

a data.frame, an object of class pltree, or a list with PlackettLuce models

...

further arguments passed to methods

bootstrap

logical, to run a bayes bootstrap on object

normalize

logical, to normalize values to sum to 1

values

an index in object with the values to compute regret

items

an index in object for the different items

group

an index in object for the different scenarios

Value

A data frame with regret estimates

items

the item names

worth

the worth parameters

regret

the squared regret

worst_regret

the worst regret

References

Loomes G. & Sugden R. (1982). The Economic Journal, 92(368), 805. doi:10.2307/2232669

Bleichrodt H. & Wakker P. P. (2015). The Economic Journal, 125(583), 493–532. doi:10.1111/ecoj.12200

Author

Jacob van Etten and Kauê de Sousa

Examples

# Case 1 ####
library("PlackettLuce")
data("breadwheat", package = "gosset")

# convert the tricot rankings from breadwheat data
# into a object of class 'grouped_rankings'

G = rank_tricot(breadwheat,
                 items = c("variety_a","variety_b","variety_c"),
                 input = c("overall_best","overall_worst"),
                 group = TRUE)

# combine grouped rankings with temperature indices
mydata = cbind(G, breadwheat[c("lon","lat")])

# fit a pltree model using geographic data
mod = pltree(G ~ ., data = mydata)

regret(mod)
#>         items worth worst_regret regret
#>         <chr> <dbl>        <dbl>  <dbl>
#> 1:      K9107 0.362        0.000  0.000
#> 2:      CSW18 0.192        0.039  0.018
#> 3:     PBW343 0.155        0.054  0.027
#> 4:      DBW17 0.064        0.061  0.056
#> 5:  DPW621-50 0.041        0.068  0.066
#> ---                                    
#> 12:    PBW502 0.014        0.071  0.077
#> 13:    HI1563 0.014        0.071  0.077
#> 14:    HD2824 0.013        0.071  0.078
#> 15:     WR544 0.011        0.073  0.078
#> 16:     K0307 0.010        0.073  0.079

# Case 2 ####
# list of PlackettLuce models
R = matrix(c(1, 2, 3, 0,
              4, 1, 2, 3,
              2, 1, 3, 4,
              1, 2, 3, 0,
              2, 1, 3, 0,
              1, 0, 3, 2), nrow = 6, byrow = TRUE)
colnames(R) = c("apple", "banana", "orange", "pear")

mod1 = PlackettLuce(R)

R2 = matrix(c(1, 2, 0, 3,
               2, 1, 0, 3,
               2, 1, 0, 3,
               1, 2, 0, 3,
               2, 1, 0, 3,
               1, 3, 4, 2), nrow = 6, byrow = TRUE)
colnames(R2) = c("apple", "banana", "orange", "pear")

mod2 = PlackettLuce(R2)

mod = list(mod1, mod2)

regret(mod, n1 = 500)
#>     items worth worst_regret regret
#>     <chr> <dbl>        <dbl>  <dbl>
#> 1: banana 0.475        0.081  0.007
#> 2:  apple 0.387        0.210  0.046
#> 3:   pear 0.078        0.348  0.455
#> 4: orange 0.060        0.361  0.492